test reorganize coverage suites
