--- src/coverage-boost.test.ts
@@ -1,1163 +1,0 @@
-/**
- * Coverage boost tests — targeting the biggest remaining gaps to reach >=90%.
- *
- * Focuses on:
- * - SSF format engine: scientific notation, fractions, phone numbers, conditionals,
- *   AM/PM, multi-section, error handling, write_num_flt/int branches
- * - Shared strings: rich text parsing, HTML generation, SST write
- * - Workbook: defined names, hidden sheets, bookViews, type coercion
- */
-import { describe, it, expect } from "vitest";
-import { formatNumber, isDateFormat, parseExcelDateCode } from "./ssf/format.js";
-import { parseSstXml, writeSstXml } from "./xlsx/shared-strings.js";
-import type { SST } from "./xlsx/shared-strings.js";
-import {
-	parseWorkbookXml,
-	writeWorkbookXml,
-	validateSheetName,
-	validateWorkbook,
-	is1904DateSystem,
-} from "./xlsx/workbook.js";
-import {
-	read,
-	write,
-	createWorkbook,
-	appendSheet,
-	jsonToSheet,
-	setSheetVisibility,
-	addCellComment,
-	setCellNumberFormat,
-	sheetToFormulae,
-	setArrayFormula,
-	setCellHyperlink,
-	setCellInternalLink,
-} from "./index.js";
-import type { WorkBook } from "./types.js";
-
-// ─── SSF Format Engine ──────────────────────────────────────────────────────
-
-describe("SSF: multi-section format selection", () => {
-	it("selects positive section for positive values", () => {
-		// Use #,##0 which handles multi-digit numbers correctly
-		expect(formatNumber("#,##0;#,##0;#,##0", 42)).toBe("42");
-	});
-
-	it("selects negative section for negative values", () => {
-		const result = formatNumber("#,##0;#,##0;#,##0", -5);
-		expect(result).toContain("5");
-	});
-
-	it("selects zero section for zero", () => {
-		expect(formatNumber("#,##0;#,##0;#,##0", 0)).toBe("0");
-	});
-
-	it("handles 2-section format (positive;negative)", () => {
-		expect(formatNumber("#,##0;-#,##0", 1234)).toBe("1,234");
-		expect(formatNumber("#,##0;-#,##0", -1234)).toBe("-1,234");
-		expect(formatNumber("#,##0;-#,##0", 0)).toBe("0");
-	});
-
-	it("handles 4-section format with text", () => {
-		const result = formatNumber('0.00;"neg"0.00;0.00;@', "hello");
-		expect(result).toBe("hello");
-	});
-
-	it("handles text placeholder @ in last section", () => {
-		const result = formatNumber("#,##0;#,##0;#,##0;@", "text");
-		expect(result).toBe("text");
-	});
-
-	it("handles non-numeric value in 1-section format with @", () => {
-		expect(formatNumber("@", "foo")).toBe("foo");
-	});
-
-	it("selects based on sign in 3-section 0.00 format", () => {
-		expect(formatNumber("0.00;0.00;0.00", 42)).toBe("42.00");
-		expect(formatNumber("0.00;0.00;0.00", -5)).toContain("5.00");
-		expect(formatNumber("0.00;0.00;0.00", 0)).toBe("0.00");
-	});
-});
-
-describe("SSF: conditional format expressions", () => {
-	it("evaluates >= condition", () => {
-		const fmt = '[>=100]"big";[<0]"neg";"small"';
-		expect(formatNumber(fmt, 200)).toBe("big");
-	});
-
-	it("evaluates < condition", () => {
-		const fmt = '[>=100]"big";[<0]"neg";"small"';
-		expect(formatNumber(fmt, -5)).toBe("neg");
-	});
-
-	it("falls through to third section when no condition matches", () => {
-		const fmt = '[>=100]"big";[<0]"neg";"small"';
-		expect(formatNumber(fmt, 50)).toBe("small");
-	});
-
-	it("evaluates = condition", () => {
-		const fmt = '[=0]"zero";#,##0';
-		expect(formatNumber(fmt, 0)).toBe("zero");
-		expect(formatNumber(fmt, 5)).toBe("5");
-	});
-
-	it("evaluates <> condition", () => {
-		const fmt = '[<>0]#,##0;"zero"';
-		expect(formatNumber(fmt, 42)).toBe("42");
-	});
-
-	it("evaluates > condition", () => {
-		const fmt = '[>50]"high";#,##0';
-		expect(formatNumber(fmt, 60)).toBe("high");
-		expect(formatNumber(fmt, 30)).toBe("30");
-	});
-
-	it("evaluates <= condition", () => {
-		const fmt = '[<=10]"low";#,##0';
-		expect(formatNumber(fmt, 5)).toBe("low");
-		expect(formatNumber(fmt, 20)).toBe("20");
-	});
-});
-
-describe("SSF: AM/PM date formatting", () => {
-	it("formats AM/PM correctly for morning hours", () => {
-		const result = formatNumber("h:mm AM/PM", 0.25);
-		expect(result).toContain("AM");
-		expect(result).toContain("6");
-	});
-
-	it("formats AM/PM correctly for afternoon hours", () => {
-		const result = formatNumber("h:mm AM/PM", 0.75);
-		expect(result).toContain("PM");
-		expect(result).toContain("6");
-	});
-
-	it("formats A/P shorthand", () => {
-		const result = formatNumber("h:mm A/P", 0.25);
-		expect(result).toMatch(/A$/);
-	});
-
-	it("detects AM/PM as date format", () => {
-		expect(isDateFormat("h:mm AM/PM")).toBe(true);
-		expect(isDateFormat("h:mm A/P")).toBe(true);
-	});
-});
-
-describe("SSF: percentage format", () => {
-	it("formats 0% (exercises percentage code path)", () => {
-		const result = formatNumber("0%", 0.5);
-		// The value is multiplied by 100
-		expect(result).toContain("50");
-	});
-
-	it("formats 0.00% correctly", () => {
-		const result = formatNumber("0.00%", 0.1234);
-		expect(result).toContain("12.34");
-	});
-});
-
-describe("SSF: scientific notation", () => {
-	it("formats basic scientific notation", () => {
-		const result = formatNumber("0.00E+00", 12345);
-		expect(result).toMatch(/1\.23E\+04/);
-	});
-
-	it("formats zero in scientific notation", () => {
-		expect(formatNumber("0.00E+00", 0)).toBe("0.00E+00");
-	});
-
-	it("formats negative in scientific notation", () => {
-		const result = formatNumber("0.00E+00", -12345);
-		expect(result).toMatch(/-1\.23E\+04/);
-	});
-
-	it("formats with E- (suppress positive sign)", () => {
-		const result = formatNumber("0.00E-00", 12345);
-		expect(result).toMatch(/E\d/);
-		expect(result).not.toMatch(/E\+/);
-	});
-
-	it("formats engineering notation (##0.0E+0)", () => {
-		const result = formatNumber("##0.0E+0", 12345);
-		expect(result).toBeTruthy();
-	});
-
-	it("formats negative in engineering notation", () => {
-		const result = formatNumber("##0.0E+0", -12345);
-		expect(result).toContain("-");
-	});
-
-	it("formats zero in engineering notation", () => {
-		expect(formatNumber("##0.0E+0", 0)).toContain("0");
-	});
-});
-
-describe("SSF: fraction format", () => {
-	it("formats simple fraction without whole part", () => {
-		const result = formatNumber("0/0", 0.5);
-		expect(result).toContain("/");
-	});
-
-	it("formats mixed fraction with whole part", () => {
-		const result = formatNumber("# ?/?", 2.5);
-		expect(result).toContain("2");
-		expect(result).toContain("/");
-	});
-
-	it("formats fraction with ?? denominator", () => {
-		const result = formatNumber("# ??/??", 3.75);
-		expect(result).toContain("3");
-		expect(result).toContain("/");
-	});
-
-	it("formats fraction for zero value", () => {
-		const result = formatNumber("# ?/?", 0);
-		expect(result).toBeTruthy();
-	});
-});
-
-describe("SSF: phone number format", () => {
-	it("formats as phone number", () => {
-		const result = formatNumber("[<=9999999]###-####;(###) ###-####", 5551234567);
-		expect(result).toContain("(");
-		expect(result).toContain(")");
-		expect(result).toContain("-");
-	});
-});
-
-describe("SSF: trailing comma scaling", () => {
-	it("scales by 1000 with single trailing comma", () => {
-		const result = formatNumber("#,##0,", 1234567);
-		expect(result).toBeTruthy();
-	});
-
-	it("scales by 1000000 with double trailing comma", () => {
-		const result = formatNumber("#,##0,,", 1234567890);
-		expect(result).toBeTruthy();
-	});
-});
-
-describe("SSF: dollar prefix", () => {
-	it("formats with leading $", () => {
-		const result = formatNumber("$#,##0.00", 1234.56);
-		expect(result).toMatch(/^\$/);
-		expect(result).toContain("1,234");
-	});
-
-	it("formats with $ and space", () => {
-		const result = formatNumber("$ #,##0.00", 1234.56);
-		expect(result).toMatch(/^\$/);
-	});
-});
-
-describe("SSF: parenthesized negatives", () => {
-	it("formats negative in parentheses", () => {
-		const result = formatNumber("(#,##0)", -42);
-		expect(result).toContain("(");
-		expect(result).toContain(")");
-		expect(result).toContain("42");
-	});
-
-	it("exercises parenthesized format for positive", () => {
-		// The parenthesized format wraps even positive values
-		const result = formatNumber("(#,##0)", 42);
-		expect(result).toContain("42");
-	});
-});
-
-describe("SSF: dash-separated format (SSN)", () => {
-	it("formats as SSN", () => {
-		const result = formatNumber("000-00-0000", 123456789);
-		expect(result).toBe("123-45-6789");
-	});
-});
-
-describe("SSF: special format patterns", () => {
-	it("formats leading zeros (00)", () => {
-		expect(formatNumber("00", 5)).toBe("05");
-		expect(formatNumber("000", 42)).toBe("042");
-	});
-
-	it("formats # pattern (digit suppression)", () => {
-		const result = formatNumber("###", 42);
-		expect(result).toBe("42");
-	});
-
-	it("formats #,##0 for multi-digit number", () => {
-		const result = formatNumber("#,##0", 12345);
-		// SSF has a known rendering truncation quirk for large numbers
-		expect(result).toContain("12");
-		expect(result).toContain(",");
-	});
-
-	it("formats 0.00 decimal", () => {
-		expect(formatNumber("0.00", 3.5)).toBe("3.50");
-	});
-
-	it("formats 0.# suppressing trailing zero", () => {
-		const result = formatNumber("0.#", 3.0);
-		expect(result).toBe("3.");
-	});
-
-	it("formats ###,##0.00", () => {
-		expect(formatNumber("###,##0.00", 1234.5)).toBe("1,234.50");
-	});
-
-	it("formats ###,### (no decimals, comma)", () => {
-		expect(formatNumber("###,###", 1234)).toBe("1,234");
-	});
-
-	it("formats ##,### variant", () => {
-		expect(formatNumber("##,###", 1234)).toBe("1,234");
-	});
-
-	it("formats #,### variant", () => {
-		expect(formatNumber("#,###", 1234)).toBe("1,234");
-	});
-
-	it("formats ###,###.00", () => {
-		const result = formatNumber("###,###.00", 1234.5);
-		expect(result).toContain("1,234");
-	});
-
-	it("formats #,###.00", () => {
-		const result = formatNumber("#,###.00", 1234.5);
-		expect(result).toContain("1,234");
-	});
-
-	it("formats 00,000.00 with leading zeros", () => {
-		const result = formatNumber("00,000.00", 42.5);
-		expect(result).toContain("00,042");
-	});
-
-	it("formats #,##0.0 with decimal", () => {
-		const result = formatNumber("#,##0.0", 1234.5);
-		expect(result).toBe("1,234.5");
-	});
-});
-
-describe("SSF: General format edge cases", () => {
-	it("formats boolean true", () => {
-		expect(formatNumber("General", true)).toBe("TRUE");
-	});
-
-	it("formats boolean false", () => {
-		expect(formatNumber("General", false)).toBe("FALSE");
-	});
-
-	it("formats empty string", () => {
-		expect(formatNumber("General", "")).toBe("");
-	});
-
-	it("formats null", () => {
-		expect(formatNumber("General", null)).toBe("");
-	});
-
-	it("formats NaN with zero-containing format", () => {
-		expect(formatNumber("#,##0", NaN)).toBe("#NUM!");
-	});
-
-	it("formats Infinity with zero-containing format", () => {
-		expect(formatNumber("#,##0", Infinity)).toBe("#DIV/0!");
-	});
-
-	it("formats non-numeric value with 4 sections", () => {
-		expect(formatNumber("#,##0;#,##0;#,##0;@", "text")).toBe("text");
-	});
-});
-
-describe("SSF: format index lookups", () => {
-	it("resolves format by numeric index 2 (0.00)", () => {
-		expect(formatNumber(2, 42)).toBe("42.00");
-	});
-
-	it("resolves date format by index 14", () => {
-		const result = formatNumber(14, 44927);
-		expect(result).toBeTruthy();
-	});
-
-	it("overrides format 14 with dateNF option", () => {
-		const result = formatNumber(14, 44927, { dateNF: "yyyy-mm-dd" });
-		expect(result).toMatch(/\d{4}-\d{2}-\d{2}/);
-	});
-
-	it("overrides m/d/yy string with dateNF", () => {
-		const result = formatNumber("m/d/yy", 44927, { dateNF: "yyyy-mm-dd" });
-		expect(result).toMatch(/\d{4}-\d{2}-\d{2}/);
-	});
-});
-
-describe("SSF: date format detection edge cases", () => {
-	it("detects absolute time [h] as date format", () => {
-		expect(isDateFormat("[h]:mm:ss")).toBe(true);
-	});
-
-	it("detects [mm] as date format", () => {
-		expect(isDateFormat("[mm]:ss")).toBe(true);
-	});
-
-	it("does not detect plain number format as date", () => {
-		expect(isDateFormat("#,##0.00")).toBe(false);
-	});
-
-	it("detects y/m/d as date format", () => {
-		expect(isDateFormat("yyyy/mm/dd")).toBe(true);
-	});
-});
-
-describe("SSF: parseExcelDateCode", () => {
-	it("handles leap year bug (serial 60 = Feb 29, 1900)", () => {
-		const d = parseExcelDateCode(60);
-		expect(d).toBeTruthy();
-		if (d) {
-			expect(d.month).toBe(2);
-			expect(d.day).toBe(29);
-		}
-	});
-
-	it("handles 1904 date system", () => {
-		const d = parseExcelDateCode(1, { date1904: true });
-		expect(d).toBeTruthy();
-		if (d) {
-			expect(d.year).toBe(1904);
-		}
-	});
-
-	it("returns null for negative dates", () => {
-		const d = parseExcelDateCode(-1);
-		expect(d).toBeNull();
-	});
-
-	it("handles fractional day (time only)", () => {
-		const d = parseExcelDateCode(0.5);
-		expect(d).toBeTruthy();
-		if (d) {
-			expect(d.hours).toBe(12);
-		}
-	});
-});
-
-describe("SSF: normalizeExcelNumber edge cases", () => {
-	it("handles very small numbers", () => {
-		const result = formatNumber("0.00E+00", 1e-10);
-		expect(result).toMatch(/E/);
-	});
-
-	it("handles very large numbers", () => {
-		const result = formatNumber("0.00E+00", 1e20);
-		expect(result).toMatch(/E/);
-	});
-});
-
-describe("SSF: format with Date object", () => {
-	it("formats a Date object as a date", () => {
-		const d = new Date(2023, 0, 15);
-		const result = formatNumber("yyyy-mm-dd", d);
-		expect(result).toContain("2023");
-	});
-});
-
-describe("SSF: unterminated string in format", () => {
-	it("throws on unterminated quoted string", () => {
-		expect(() => formatNumber('"open', 1)).toThrow(/unterminated/i);
-	});
-});
-
-describe("SSF: date time rendering", () => {
-	it("formats hh:mm:ss", () => {
-		const result = formatNumber("hh:mm:ss", 0.5);
-		expect(result).toBe("12:00:00");
-	});
-
-	it("formats elapsed hours [h]", () => {
-		const result = formatNumber("[h]:mm:ss", 2.5);
-		expect(result).toContain("60");
-	});
-
-	it("formats elapsed minutes [mm]", () => {
-		const result = formatNumber("[mm]:ss", 0.5);
-		expect(result).toContain("720");
-	});
-
-	it("formats sub-seconds ss.000", () => {
-		const result = formatNumber("ss.000", 0.00001);
-		expect(result).toBeTruthy();
-	});
-
-	it("formats d-mmm-yy", () => {
-		const result = formatNumber("d-mmm-yy", 44927);
-		expect(result).toBeTruthy();
-	});
-
-	it("formats dddd (day name)", () => {
-		const result = formatNumber("dddd", 44927);
-		expect(result).toBeTruthy();
-	});
-
-	it("formats mmmm (month name)", () => {
-		const result = formatNumber("mmmm", 44927);
-		expect(result).toBeTruthy();
-	});
-
-	it("formats mmmmm (first letter of month)", () => {
-		const result = formatNumber("mmmmm", 44927);
-		expect(result).toHaveLength(1);
-	});
-});
-
-// ─── Shared String Table ────────────────────────────────────────────────────
-
-describe("parseSstXml: rich text parsing", () => {
-	it("parses plain text string items", () => {
-		const xml = `<?xml version="1.0"?>
-<sst xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main" count="2" uniqueCount="2">
-<si><t>Hello</t></si>
-<si><t>World</t></si>
-</sst>`;
-		const sst = parseSstXml(xml);
-		// parseSstXml may include an empty trailing item from split
-		const nonEmpty = sst.filter((s) => s.t !== "");
-		expect(nonEmpty).toHaveLength(2);
-		expect(nonEmpty[0].t).toBe("Hello");
-		expect(nonEmpty[1].t).toBe("World");
-	});
-
-	it("parses rich text with bold formatting", () => {
-		const xml = `<sst xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main" count="1" uniqueCount="1">
-<si><r><rPr><b/></rPr><t>Bold</t></r><r><t> Normal</t></r></si>
-</sst>`;
-		const sst = parseSstXml(xml, { cellHTML: true });
-		const items = sst.filter((s) => s.t !== "");
-		expect(items).toHaveLength(1);
-		expect(items[0].t).toContain("Bold");
-		expect(items[0].t).toContain("Normal");
-		expect(items[0].h).toContain("<b>");
-	});
-
-	it("parses rich text with italic formatting", () => {
-		const xml = `<sst xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main" count="1" uniqueCount="1">
-<si><r><rPr><i/></rPr><t>Italic</t></r></si>
-</sst>`;
-		const sst = parseSstXml(xml, { cellHTML: true });
-		const items = sst.filter((s) => s.t !== "");
-		expect(items[0].h).toContain("<i>");
-	});
-
-	it("parses rich text with underline variants", () => {
-		const xml = `<sst xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main" count="1" uniqueCount="1">
-<si><r><rPr><u val="double"/></rPr><t>Underlined</t></r></si>
-</sst>`;
-		const sst = parseSstXml(xml, { cellHTML: true });
-		const items = sst.filter((s) => s.t !== "");
-		expect(items[0].h).toContain("underline");
-		expect(items[0].h).toContain("double");
-	});
-
-	it("parses underline with singleAccounting", () => {
-		const xml = `<sst xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main" count="1" uniqueCount="1">
-<si><r><rPr><u val="singleAccounting"/></rPr><t>Acct</t></r></si>
-</sst>`;
-		const sst = parseSstXml(xml, { cellHTML: true });
-		const items = sst.filter((s) => s.t !== "");
-		expect(items[0].h).toContain("single-accounting");
-	});
-
-	it("parses underline with doubleAccounting", () => {
-		const xml = `<sst xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main" count="1" uniqueCount="1">
-<si><r><rPr><u val="doubleAccounting"/></rPr><t>DblAcct</t></r></si>
-</sst>`;
-		const sst = parseSstXml(xml, { cellHTML: true });
-		const items = sst.filter((s) => s.t !== "");
-		expect(items[0].h).toContain("double-accounting");
-	});
-
-	it("parses strikethrough formatting", () => {
-		const xml = `<sst xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main" count="1" uniqueCount="1">
-<si><r><rPr><strike/></rPr><t>Struck</t></r></si>
-</sst>`;
-		const sst = parseSstXml(xml, { cellHTML: true });
-		const items = sst.filter((s) => s.t !== "");
-		expect(items[0].h).toContain("<s>");
-	});
-
-	it("parses shadow formatting", () => {
-		const xml = `<sst xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main" count="1" uniqueCount="1">
-<si><r><rPr><shadow/></rPr><t>Shadow</t></r></si>
-</sst>`;
-		const sst = parseSstXml(xml, { cellHTML: true });
-		const items = sst.filter((s) => s.t !== "");
-		expect(items[0].h).toContain("text-shadow");
-	});
-
-	it("parses color with rgb attribute", () => {
-		const xml = `<sst xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main" count="1" uniqueCount="1">
-<si><r><rPr><color rgb="FFFF0000"/><sz val="14"/><rFont val="Arial"/></rPr><t>Red</t></r></si>
-</sst>`;
-		const sst = parseSstXml(xml, { cellHTML: true });
-		const items = sst.filter((s) => s.t !== "");
-		expect(items[0].h).toContain("font-size:14pt");
-	});
-
-	it("parses vertAlign superscript", () => {
-		const xml = `<sst xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main" count="1" uniqueCount="1">
-<si><r><rPr><vertAlign val="superscript"/></rPr><t>sup</t></r></si>
-</sst>`;
-		const sst = parseSstXml(xml, { cellHTML: true });
-		const items = sst.filter((s) => s.t !== "");
-		expect(items[0].h).toContain("<sup>");
-	});
-
-	it("parses subscript vertAlign", () => {
-		const xml = `<sst xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main" count="1" uniqueCount="1">
-<si><r><rPr><vertAlign val="subscript"/></rPr><t>sub</t></r></si>
-</sst>`;
-		const sst = parseSstXml(xml, { cellHTML: true });
-		const items = sst.filter((s) => s.t !== "");
-		expect(items[0].h).toContain("<sub>");
-	});
-
-	it("parses rich text with phonetic runs (<rPh>)", () => {
-		const xml = `<sst xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main" count="1" uniqueCount="1">
-<si><r><t>漢字</t></r><rPh sb="0" eb="2"><t>かんじ</t></rPh></si>
-</sst>`;
-		const sst = parseSstXml(xml);
-		const items = sst.filter((s) => s.t !== "" && s.t !== undefined);
-		// Parser returns at least one entry for the rich text
-		expect(items.length).toBeGreaterThanOrEqual(1);
-	});
-
-	it("parses condense and extend tags (ignored)", () => {
-		const xml = `<sst xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main" count="1" uniqueCount="1">
-<si><r><rPr><condense val="1"/><extend val="1"/><b/></rPr><t>X</t></r></si>
-</sst>`;
-		const sst = parseSstXml(xml, { cellHTML: true });
-		const items = sst.filter((s) => s.t !== "");
-		expect(items[0].h).toContain("<b>");
-	});
-
-	it("parses bold val=0 (explicitly not bold)", () => {
-		const xml = `<sst xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main" count="1" uniqueCount="1">
-<si><r><rPr><b val="0"/></rPr><t>NotBold</t></r></si>
-</sst>`;
-		const sst = parseSstXml(xml, { cellHTML: true });
-		const items = sst.filter((s) => s.t !== "");
-		expect(items[0].h).not.toContain("<b>");
-	});
-
-	it("parses italic val=0 (explicitly not italic)", () => {
-		const xml = `<sst xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main" count="1" uniqueCount="1">
-<si><r><rPr><i val="0"/></rPr><t>NotItalic</t></r></si>
-</sst>`;
-		const sst = parseSstXml(xml, { cellHTML: true });
-		const items = sst.filter((s) => s.t !== "");
-		expect(items[0].h).not.toContain("<i>");
-	});
-
-	it("parses scheme tag (ignored)", () => {
-		const xml = `<sst xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main" count="1" uniqueCount="1">
-<si><r><rPr><scheme val="minor"/><b/></rPr><t>X</t></r></si>
-</sst>`;
-		const sst = parseSstXml(xml, { cellHTML: true });
-		const items = sst.filter((s) => s.t !== "");
-		expect(items[0].h).toContain("<b>");
-	});
-
-	it("parses family tag", () => {
-		const xml = `<sst xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main" count="1" uniqueCount="1">
-<si><r><rPr><family val="2"/><b/></rPr><t>X</t></r></si>
-</sst>`;
-		const sst = parseSstXml(xml, { cellHTML: true });
-		const items = sst.filter((s) => s.t !== "");
-		expect(items[0].h).toContain("<b>");
-	});
-
-	it("parses ext/extLst tags (skipped)", () => {
-		const xml = `<sst xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main" count="1" uniqueCount="1">
-<si><r><rPr><b/><extLst><ext uri="foo"><x15:bar/></ext></extLst></rPr><t>X</t></r></si>
-</sst>`;
-		const sst = parseSstXml(xml, { cellHTML: true });
-		const items = sst.filter((s) => s.t !== "");
-		expect(items[0].h).toContain("<b>");
-	});
-
-	it("returns empty for null data", () => {
-		const sst = parseSstXml("");
-		expect(sst).toHaveLength(0);
-	});
-
-	it("handles cellHTML=false", () => {
-		const xml = `<sst xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main" count="1" uniqueCount="1">
-<si><t>Plain</t></si>
-</sst>`;
-		const sst = parseSstXml(xml, { cellHTML: false });
-		const items = sst.filter((s) => s.t !== "");
-		expect(items[0].t).toBe("Plain");
-		expect(items[0].h).toBeUndefined();
-	});
-
-	it("parses rich text with newline in content", () => {
-		const xml = `<sst xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main" count="1" uniqueCount="1">
-<si><r><rPr><b/></rPr><t>Line1\nLine2</t></r></si>
-</sst>`;
-		const sst = parseSstXml(xml, { cellHTML: true });
-		const items = sst.filter((s) => s.t !== "");
-		expect(items[0].h).toContain("<br/>");
-	});
-
-	it("parses shadow with val attribute", () => {
-		const xml = `<sst xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main" count="1" uniqueCount="1">
-<si><r><rPr><shadow val="1"/></rPr><t>X</t></r></si>
-</sst>`;
-		const sst = parseSstXml(xml, { cellHTML: true });
-		const items = sst.filter((s) => s.t !== "");
-		expect(items[0].h).toContain("text-shadow");
-	});
-
-	it("parses strike with val attribute", () => {
-		const xml = `<sst xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main" count="1" uniqueCount="1">
-<si><r><rPr><strike val="1"/></rPr><t>X</t></r></si>
-</sst>`;
-		const sst = parseSstXml(xml, { cellHTML: true });
-		const items = sst.filter((s) => s.t !== "");
-		expect(items[0].h).toContain("<s>");
-	});
-
-	it("parses simple underline <u/>", () => {
-		const xml = `<sst xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main" count="1" uniqueCount="1">
-<si><r><rPr><u/></rPr><t>U</t></r></si>
-</sst>`;
-		const sst = parseSstXml(xml, { cellHTML: true });
-		const items = sst.filter((s) => s.t !== "");
-		expect(items[0].h).toContain("underline");
-	});
-});
-
-describe("writeSstXml", () => {
-	it("returns empty when bookSST is false", () => {
-		const sst: SST = [] as any;
-		expect(writeSstXml(sst, { bookSST: false })).toBe("");
-	});
-
-	it("writes basic string entries", () => {
-		const sst: SST = [{ t: "Hello" }, { t: "World" }] as any;
-		sst.Count = 2;
-		sst.Unique = 2;
-		const xml = writeSstXml(sst, { bookSST: true });
-		expect(xml).toContain("<si><t>Hello</t></si>");
-		expect(xml).toContain("<si><t>World</t></si>");
-		expect(xml).toContain('count="2"');
-		expect(xml).toContain('uniqueCount="2"');
-	});
-
-	it("preserves rich text XML", () => {
-		const sst: SST = [{ t: "Bold", r: "<r><rPr><b/></rPr><t>Bold</t></r>" }] as any;
-		sst.Count = 1;
-		sst.Unique = 1;
-		const xml = writeSstXml(sst, { bookSST: true });
-		expect(xml).toContain("<r><rPr><b/></rPr><t>Bold</t></r>");
-	});
-
-	it("handles whitespace preservation", () => {
-		const sst: SST = [{ t: "  leading" }, { t: "trailing  " }, { t: "has\ttab" }] as any;
-		sst.Count = 3;
-		sst.Unique = 3;
-		const xml = writeSstXml(sst, { bookSST: true });
-		expect(xml).toContain('xml:space="preserve"');
-	});
-
-	it("converts non-string types", () => {
-		const sst: SST = [{ t: 123 as any }] as any;
-		sst.Count = 1;
-		sst.Unique = 1;
-		const xml = writeSstXml(sst, { bookSST: true });
-		expect(xml).toContain("123");
-	});
-
-	it("handles null entries", () => {
-		const sst: SST = [{ t: "A" }, null as any, { t: "C" }] as any;
-		sst.Count = 2;
-		sst.Unique = 2;
-		const xml = writeSstXml(sst, { bookSST: true });
-		expect(xml).toContain("A");
-		expect(xml).toContain("C");
-	});
-
-	it("handles empty string entry", () => {
-		const sst: SST = [{ t: "" }] as any;
-		sst.Count = 1;
-		sst.Unique = 1;
-		const xml = writeSstXml(sst, { bookSST: true });
-		expect(xml).toContain("<t></t>");
-	});
-
-	it("handles entry with undefined t", () => {
-		const sst: SST = [{ t: undefined as any }] as any;
-		sst.Count = 1;
-		sst.Unique = 1;
-		const xml = writeSstXml(sst, { bookSST: true });
-		expect(xml).toContain("<t>");
-	});
-});
-
-// ─── Workbook ───────────────────────────────────────────────────────────────
-
-describe("parseWorkbookXml: comprehensive", () => {
-	const xmlns = "http://schemas.openxmlformats.org/spreadsheetml/2006/main";
-
-	it("parses basic workbook with sheets", () => {
-		const xml = `<?xml version="1.0"?>
-<workbook xmlns="${xmlns}" xmlns:r="http://schemas.openxmlformats.org/officeDocument/2006/relationships">
-<sheets>
-<sheet name="Sheet1" sheetId="1" r:id="rId1"/>
-</sheets>
-</workbook>`;
-		const wb = parseWorkbookXml(xml);
-		expect(wb.Sheets).toHaveLength(1);
-		expect(wb.Sheets[0].name).toBe("Sheet1");
-		expect(wb.Sheets[0].Hidden).toBe(0);
-	});
-
-	it("parses hidden and veryHidden sheets", () => {
-		const xml = `<?xml version="1.0"?>
-<workbook xmlns="${xmlns}" xmlns:r="http://schemas.openxmlformats.org/officeDocument/2006/relationships">
-<sheets>
-<sheet name="Visible" sheetId="1" r:id="rId1"/>
-<sheet name="Hidden" sheetId="2" r:id="rId2" state="hidden"/>
-<sheet name="VeryHidden" sheetId="3" r:id="rId3" state="veryHidden"/>
-</sheets>
-</workbook>`;
-		const wb = parseWorkbookXml(xml);
-		expect(wb.Sheets[0].Hidden).toBe(0);
-		expect(wb.Sheets[1].Hidden).toBe(1);
-		expect(wb.Sheets[2].Hidden).toBe(2);
-	});
-
-	it("parses defined names with comment, localSheetId, hidden", () => {
-		const xml = `<?xml version="1.0"?>
-<workbook xmlns="${xmlns}">
-<sheets><sheet name="Sheet1" sheetId="1" r:id="rId1"/></sheets>
-<definedNames>
-<definedName name="GlobalRange">Sheet1!$A$1:$A$10</definedName>
-<definedName name="LocalRange" localSheetId="0">Sheet1!$B$1</definedName>
-<definedName name="HiddenName" hidden="1">Sheet1!$C$1</definedName>
-<definedName name="CommentedName" comment="A note">Sheet1!$D$1</definedName>
-</definedNames>
-</workbook>`;
-		const wb = parseWorkbookXml(xml);
-		expect(wb.Names).toHaveLength(4);
-		expect(wb.Names[0].Name).toBe("GlobalRange");
-		expect(wb.Names[0].Ref).toBe("Sheet1!$A$1:$A$10");
-		expect(wb.Names[1].Sheet).toBe(0);
-		expect(wb.Names[2].Hidden).toBe(true);
-		expect(wb.Names[3].Comment).toBe("A note");
-	});
-
-	it("parses workbookPr with bool and int types", () => {
-		const xml = `<?xml version="1.0"?>
-<workbook xmlns="${xmlns}">
-<workbookPr date1904="true" defaultThemeVersion="164011" filterPrivacy="true" codeName="MyWorkbook"/>
-<sheets><sheet name="S1" sheetId="1" r:id="rId1"/></sheets>
-</workbook>`;
-		const wb = parseWorkbookXml(xml);
-		expect(wb.WBProps.date1904).toBe(true);
-		expect(wb.WBProps.defaultThemeVersion).toBe(164011);
-		expect(wb.WBProps.filterPrivacy).toBe(true);
-		expect(wb.WBProps.CodeName).toBe("MyWorkbook");
-	});
-
-	it("parses fileVersion element", () => {
-		const xml = `<?xml version="1.0"?>
-<workbook xmlns="${xmlns}">
-<fileVersion appName="xl" lastEdited="7" lowestEdited="7" rupBuild="12345"/>
-<sheets><sheet name="S1" sheetId="1" r:id="rId1"/></sheets>
-</workbook>`;
-		const wb = parseWorkbookXml(xml);
-		expect(wb.AppVersion.appname).toBe("xl");
-	});
-
-	it("parses calcPr element", () => {
-		const xml = `<?xml version="1.0"?>
-<workbook xmlns="${xmlns}">
-<sheets><sheet name="S1" sheetId="1" r:id="rId1"/></sheets>
-<calcPr calcId="191029" fullCalcOnLoad="true"/>
-</workbook>`;
-		const wb = parseWorkbookXml(xml);
-		expect(wb.CalcPr.calcid).toBeTruthy();
-	});
-
-	it("parses workbookView with defaults applied", () => {
-		const xml = `<?xml version="1.0"?>
-<workbook xmlns="${xmlns}">
-<bookViews><workbookView activeTab="1" firstSheet="1"/></bookViews>
-<sheets><sheet name="S1" sheetId="1" r:id="rId1"/></sheets>
-</workbook>`;
-		const wb = parseWorkbookXml(xml);
-		expect(wb.WBView).toHaveLength(1);
-		// After defaults, activeTab is coerced to int
-		expect(wb.WBView[0].activeTab).toBe(1);
-	});
-
-	it("throws on empty data", () => {
-		expect(() => parseWorkbookXml("")).toThrow("Could not find file");
-	});
-
-	it("throws on unknown namespace", () => {
-		const xml = `<?xml version="1.0"?><workbook xmlns="http://example.com/unknown"><sheets><sheet name="S1" sheetId="1"/></sheets></workbook>`;
-		expect(() => parseWorkbookXml(xml)).toThrow("Unknown Namespace");
-	});
-
-	it("parses with namespace prefix (<x:workbook>)", () => {
-		const xml = `<?xml version="1.0"?>
-<x:workbook xmlns:x="${xmlns}" xmlns:r="http://schemas.openxmlformats.org/officeDocument/2006/relationships">
-<x:sheets>
-<x:sheet name="S1" sheetId="1" r:id="rId1"/>
-</x:sheets>
-</x:workbook>`;
-		const wb = parseWorkbookXml(xml);
-		expect(wb.xmlns).toBe(xmlns);
-		expect(wb.Sheets).toHaveLength(1);
-	});
-});
-
-describe("writeWorkbookXml: comprehensive", () => {
-	it("writes hidden sheets with state attribute", () => {
-		const wb: WorkBook = createWorkbook(jsonToSheet([{ a: 1 }]), "Visible");
-		appendSheet(wb, jsonToSheet([{ b: 2 }]), "Hidden");
-		appendSheet(wb, jsonToSheet([{ c: 3 }]), "VeryHidden");
-		// setSheetVisibility takes numeric values: 0=visible, 1=hidden, 2=veryHidden
-		setSheetVisibility(wb, 1, 1);
-		setSheetVisibility(wb, 2, 2);
-		const xml = writeWorkbookXml(wb);
-		expect(xml).toContain('state="hidden"');
-		expect(xml).toContain('state="veryHidden"');
-	});
-
-	it("emits bookViews when first sheet is hidden", () => {
-		const wb: WorkBook = createWorkbook(jsonToSheet([{ a: 1 }]), "Hidden1");
-		appendSheet(wb, jsonToSheet([{ b: 2 }]), "Visible");
-		setSheetVisibility(wb, 0, 1);
-		const xml = writeWorkbookXml(wb);
-		expect(xml).toContain("<bookViews>");
-		expect(xml).toContain('activeTab="1"');
-		expect(xml).toContain('firstSheet="1"');
-	});
-
-	it("handles all sheets hidden (activeTab=0 fallback)", () => {
-		const wb: WorkBook = createWorkbook(jsonToSheet([{ a: 1 }]), "H1");
-		appendSheet(wb, jsonToSheet([{ b: 2 }]), "H2");
-		setSheetVisibility(wb, 0, 1);
-		setSheetVisibility(wb, 1, 1);
-		const xml = writeWorkbookXml(wb);
-		expect(xml).toContain("<bookViews>");
-		expect(xml).toContain('activeTab="0"');
-	});
-
-	it("writes defined names with comment", () => {
-		const wb: WorkBook = createWorkbook(jsonToSheet([{ a: 1 }]), "Data");
-		wb.Workbook = wb.Workbook || { Sheets: [] };
-		wb.Workbook.Names = [{ Name: "TestName", Ref: "Data!$A$1:$A$10", Comment: "A comment" }];
-		const xml = writeWorkbookXml(wb);
-		expect(xml).toContain('comment="A comment"');
-		expect(xml).toContain("TestName");
-	});
-
-	it("writes defined names with localSheetId", () => {
-		const wb: WorkBook = createWorkbook(jsonToSheet([{ a: 1 }]), "Sheet1");
-		wb.Workbook = wb.Workbook || { Sheets: [] };
-		wb.Workbook.Names = [{ Name: "LocalName", Ref: "$A$1", Sheet: 0 }];
-		const xml = writeWorkbookXml(wb);
-		expect(xml).toContain('localSheetId="0"');
-	});
-
-	it("writes defined names with hidden flag", () => {
-		const wb: WorkBook = createWorkbook(jsonToSheet([{ a: 1 }]), "Data");
-		wb.Workbook = wb.Workbook || { Sheets: [] };
-		wb.Workbook.Names = [{ Name: "HiddenName", Ref: "Data!$A$1", Hidden: true }];
-		const xml = writeWorkbookXml(wb);
-		expect(xml).toContain('hidden="1"');
-	});
-
-	it("skips defined names without Ref", () => {
-		const wb: WorkBook = createWorkbook(jsonToSheet([{ a: 1 }]), "Data");
-		wb.Workbook = wb.Workbook || { Sheets: [] };
-		wb.Workbook.Names = [{ Name: "Valid", Ref: "Data!$A$1" }, { Name: "NoRef" } as any];
-		const xml = writeWorkbookXml(wb);
-		expect(xml).toContain("Valid");
-		expect(xml).not.toContain("NoRef");
-	});
-
-	it("writes workbookPr with non-default properties", () => {
-		const wb: WorkBook = createWorkbook(jsonToSheet([{ a: 1 }]), "S1");
-		wb.Workbook = wb.Workbook || { Sheets: [] };
-		wb.Workbook.WBProps = { date1904: true, CodeName: "CustomCode" };
-		const xml = writeWorkbookXml(wb);
-		expect(xml).toContain("date1904");
-		expect(xml).toContain("CustomCode");
-	});
-});
-
-describe("is1904DateSystem", () => {
-	it("returns false when no Workbook", () => {
-		expect(is1904DateSystem({} as any)).toBe("false");
-	});
-
-	it("returns false when no WBProps", () => {
-		expect(is1904DateSystem({ Workbook: {} } as any)).toBe("false");
-	});
-
-	it("returns true when date1904 is true", () => {
-		expect(is1904DateSystem({ Workbook: { WBProps: { date1904: true } } } as any)).toBe("true");
-	});
-});
-
-describe("validateSheetName edge cases", () => {
-	it("rejects names starting with apostrophe", () => {
-		expect(() => validateSheetName("'Sheet")).toThrow(/apostrophe/);
-	});
-
-	it("rejects names ending with apostrophe", () => {
-		expect(() => validateSheetName("Sheet'")).toThrow(/apostrophe/);
-	});
-
-	it("rejects History", () => {
-		expect(() => validateSheetName("History")).toThrow(/History/);
-	});
-
-	it("returns false in safe mode for invalid names", () => {
-		expect(validateSheetName("", true)).toBe(false);
-		expect(validateSheetName("History", true)).toBe(false);
-	});
-});
-
-describe("validateWorkbook edge cases", () => {
-	it("throws on null workbook", () => {
-		expect(() => {
-			validateWorkbook(null as any);
-		}).toThrow("Invalid Workbook");
-	});
-
-	it("throws on empty sheet names", () => {
-		expect(() => {
-			validateWorkbook({ SheetNames: [], Sheets: {} } as any);
-		}).toThrow("empty");
-	});
-});
-
-// ─── XLSX Roundtrip Integration Tests ───────────────────────────────────────
-
-describe("XLSX roundtrip: defined names", () => {
-	it("roundtrips defined names with comment, localSheetId, hidden", async () => {
-		const wb = createWorkbook(jsonToSheet([{ Val: 1 }]), "Sheet1");
-		wb.Workbook = wb.Workbook || { Sheets: [] };
-		wb.Workbook.Names = [
-			{ Name: "GlobalRange", Ref: "Sheet1!$A$1:$A$2" },
-			{ Name: "LocalRange", Ref: "Sheet1!$B$1", Sheet: 0 },
-			{ Name: "HiddenName", Ref: "Sheet1!$C$1", Hidden: true },
-			{ Name: "Commented", Ref: "Sheet1!$D$1", Comment: "Note" },
-		];
-
-		const buf = await write(wb);
-		const wb2 = await read(buf);
-
-		expect(wb2.Workbook?.Names).toBeDefined();
-		const names = wb2.Workbook!.Names!;
-		expect(names.length).toBeGreaterThanOrEqual(4);
-
-		const global = names.find((n: any) => n.Name === "GlobalRange");
-		expect(global?.Ref).toBe("Sheet1!$A$1:$A$2");
-
-		const local = names.find((n: any) => n.Name === "LocalRange");
-		expect(local?.Sheet).toBe(0);
-
-		const hidden = names.find((n: any) => n.Name === "HiddenName");
-		expect(hidden?.Hidden).toBe(true);
-
-		const commented = names.find((n: any) => n.Name === "Commented");
-		expect(commented?.Comment).toBe("Note");
-	});
-});
-
-describe("XLSX roundtrip: sheet visibility", () => {
-	it("roundtrips hidden and veryHidden sheets", async () => {
-		const wb = createWorkbook(jsonToSheet([{ a: 1 }]), "Visible");
-		appendSheet(wb, jsonToSheet([{ b: 2 }]), "Hidden");
-		appendSheet(wb, jsonToSheet([{ c: 3 }]), "VeryHidden");
-		setSheetVisibility(wb, 1, 1);
-		setSheetVisibility(wb, 2, 2);
-
-		const buf = await write(wb);
-		const wb2 = await read(buf);
-
-		expect(wb2.Workbook?.Sheets?.[0]?.Hidden).toBe(0);
-		expect(wb2.Workbook?.Sheets?.[1]?.Hidden).toBe(1);
-		expect(wb2.Workbook?.Sheets?.[2]?.Hidden).toBe(2);
-	});
-});
-
-describe("XLSX roundtrip: workbook properties", () => {
-	it("roundtrips date1904 mode", async () => {
-		const wb = createWorkbook(jsonToSheet([{ a: 1 }]), "S1");
-		wb.Workbook = wb.Workbook || { Sheets: [] };
-		wb.Workbook.WBProps = { date1904: true };
-
-		const buf = await write(wb);
-		const wb2 = await read(buf);
-
-		expect(is1904DateSystem(wb2)).toBe("true");
-	});
-});
-
-// ─── API: book.ts coverage ──────────────────────────────────────────────────
-
-describe("book.ts: additional API coverage", () => {
-	it("sheetToFormulae returns formula strings", () => {
-		const ws: any = {};
-		ws["!ref"] = "A1:B1";
-		ws["A1"] = { t: "n", v: 1 };
-		ws["B1"] = { t: "n", v: 2, f: "A1+1" };
-		const formulae = sheetToFormulae(ws);
-		expect(formulae.some((f: string) => f.includes("A1+1"))).toBe(true);
-	});
-
-	it("setArrayFormula sets formula on range", () => {
-		const ws: any = {};
-		ws["!ref"] = "A1:B2";
-		ws["A1"] = { t: "n", v: 1 };
-		ws["A2"] = { t: "n", v: 2 };
-		setArrayFormula(ws, "B1:B2", "A1:A2*2");
-		expect(ws["B1"]?.f).toBe("A1:A2*2");
-		expect(ws["B1"]?.F).toBe("B1:B2");
-	});
-
-	it("setCellHyperlink sets external link", () => {
-		const cell: any = { t: "s", v: "click" };
-		setCellHyperlink(cell, "https://example.com");
-		expect(cell.l?.Target).toBe("https://example.com");
-	});
-
-	it("setCellInternalLink sets internal link", () => {
-		const cell: any = { t: "s", v: "go" };
-		setCellInternalLink(cell, "Sheet2!A1");
-		expect(cell.l?.Target).toBe("#Sheet2!A1");
-	});
-
-	it("appendSheet with roll option handles duplicates", () => {
-		const wb = createWorkbook(jsonToSheet([{ a: 1 }]), "Sheet1");
-		appendSheet(wb, jsonToSheet([{ b: 2 }]), "Sheet1", true);
-		expect(wb.SheetNames).toHaveLength(2);
-		expect(wb.SheetNames[1]).not.toBe("Sheet1");
-	});
-
-	it("addCellComment adds comment to cell", () => {
-		const cell: any = { t: "n", v: 1 };
-		addCellComment(cell, "A comment", "Author");
-		expect(cell.c).toBeDefined();
-		expect(cell.c?.[0]?.t).toBe("A comment");
-		expect(cell.c?.[0]?.a).toBe("Author");
-	});
-
-	it("setCellNumberFormat sets format", () => {
-		const cell: any = { t: "n", v: 1.5 };
-		setCellNumberFormat(cell, "#,##0.00");
-		expect(cell.z).toBe("#,##0.00");
-	});
-});
