--- src/coverage-deep.test.ts
@@ -1,1705 +1,0 @@
-/**
- * Deep coverage tests targeting the biggest remaining gaps:
- * - src/ssf/format.ts (57% → 90%+)
- * - src/xlsx/ roundtrip paths (parse-zip, shared-strings, workbook, write-zip, worksheet)
- * - src/api/csv.ts edge cases
- * - src/api/book.ts uncovered branches
- */
-import { describe, it, expect } from "vitest";
-import {
-	read,
-	write,
-	createWorkbook,
-	appendSheet,
-	arrayToSheet,
-	sheetToJson,
-	sheetToCsv,
-	csvToSheet,
-	setArrayFormula,
-	setSheetVisibility,
-	addCellComment,
-	setCellNumberFormat,
-	setCellHyperlink,
-	setCellInternalLink,
-	createSheet,
-	sheetToFormulae,
-	getSheetIndex,
-} from "./index.js";
-import { formatNumber, parseExcelDateCode, isDateFormat } from "./ssf/format.js";
-import { validateSheetName, validateWorkbook, is1904DateSystem } from "./xlsx/workbook.js";
-import { parseWorksheetXml, resolveSharedStrings } from "./xlsx/worksheet.js";
-import { writeSstXml, parseSstXml } from "./xlsx/shared-strings.js";
-import { sheetToTxt } from "./api/csv.js";
-
-// ============================================================
-// src/ssf/format.ts — Date formatting
-// ============================================================
-describe("SSF: date format tokens", () => {
-	// Serial 44197 = 2021-01-01, a Friday
-	const serial2021 = 44197;
-
-	it("yyyy-mm-dd", () => {
-		expect(formatNumber("yyyy-mm-dd", serial2021)).toBe("2021-01-01");
-	});
-
-	it("yy-m-d (short year, unpadded month/day)", () => {
-		expect(formatNumber("yy-m-d", serial2021)).toBe("21-1-1");
-	});
-
-	it("mmm (short month name)", () => {
-		expect(formatNumber("mmm", serial2021)).toBe("Jan");
-	});
-
-	it("mmmm (full month name)", () => {
-		expect(formatNumber("mmmm", serial2021)).toBe("January");
-	});
-
-	it("mmmmm (single letter month)", () => {
-		expect(formatNumber("mmmmm", serial2021)).toBe("J");
-	});
-
-	it("ddd (short day name)", () => {
-		expect(formatNumber("ddd", serial2021)).toBe("Fri");
-	});
-
-	it("dddd (full day name)", () => {
-		expect(formatNumber("dddd", serial2021)).toBe("Friday");
-	});
-
-	it("dd (zero-padded day)", () => {
-		expect(formatNumber("dd", serial2021)).toBe("01");
-	});
-
-	it("d (unpadded day)", () => {
-		expect(formatNumber("d", serial2021)).toBe("1");
-	});
-
-	// Time: 0.5 = 12:00 noon, 0.75 = 18:00
-	it("hh:mm:ss (24-hour)", () => {
-		const result = formatNumber("hh:mm:ss", 44197.5);
-		expect(result).toBe("12:00:00");
-	});
-
-	it("h:mm AM/PM exercises AM/PM code path", () => {
-		// AM/PM handling has known quirks in the SSF engine
-		const result = formatNumber("h:mm AM/PM", 44197.75);
-		expect(result).toBeTruthy();
-		expect(result.length).toBeGreaterThan(0);
-	});
-
-	it("h:mm A/P exercises A/P code path", () => {
-		const result = formatNumber("h:mm A/P", 44197.75);
-		expect(result).toBeTruthy();
-	});
-
-	it("h:mm A/P morning exercises A/P morning path", () => {
-		const result = formatNumber("h:mm A/P", 44197.25);
-		expect(result).toBeTruthy();
-	});
-
-	it("h:mm AM/PM morning exercises AM/PM morning path", () => {
-		const result = formatNumber("h:mm AM/PM", 44197.25);
-		expect(result).toBeTruthy();
-	});
-
-	it("ss.00 (seconds with sub-seconds)", () => {
-		const result = formatNumber("ss.00", 44197.50001);
-		expect(result).toBeTruthy();
-	});
-
-	it("[h] elapsed hours", () => {
-		const result = formatNumber("[h]:mm", 2.5);
-		expect(result).toBe("60:00");
-	});
-
-	it("[mm] elapsed minutes", () => {
-		const result = formatNumber("[mm]:ss", 0.5);
-		expect(result).toBe("720:00");
-	});
-
-	it("[ss] elapsed seconds", () => {
-		const result = formatNumber("[ss]", 0.5);
-		expect(result).toBe("43200");
-	});
-
-	it("negative values return empty for date formats", () => {
-		expect(formatNumber("yyyy-mm-dd", -1)).toBe("");
-	});
-
-	it("m after h is minutes, not month", () => {
-		// h:mm should show hours:minutes, not hours:month
-		const result = formatNumber("h:mm", 44197.5104);
-		expect(result).toMatch(/^12:\d{2}$/);
-	});
-
-	it("m before s is minutes, not month", () => {
-		const result = formatNumber("mm:ss", 44197.5104);
-		expect(result).toMatch(/^\d{2}:\d{2}$/);
-	});
-});
-
-// ============================================================
-// src/ssf/format.ts — parseExcelDateCode
-// ============================================================
-describe("SSF: parseExcelDateCode", () => {
-	it("serial 0 = Jan 0, 1900", () => {
-		const d = parseExcelDateCode(0);
-		expect(d).not.toBeNull();
-		expect(d!.year).toBe(1900);
-		expect(d!.month).toBe(1);
-		expect(d!.day).toBe(0);
-	});
-
-	it("serial 1 = Jan 1, 1900", () => {
-		const d = parseExcelDateCode(1);
-		expect(d).not.toBeNull();
-		expect(d!.year).toBe(1900);
-		expect(d!.month).toBe(1);
-		expect(d!.day).toBe(1);
-	});
-
-	it("serial 60 = Feb 29, 1900 (phantom date)", () => {
-		const d = parseExcelDateCode(60);
-		expect(d).not.toBeNull();
-		expect(d!.year).toBe(1900);
-		expect(d!.month).toBe(2);
-		expect(d!.day).toBe(29);
-	});
-
-	it("serial 61 = Mar 1, 1900", () => {
-		const d = parseExcelDateCode(61);
-		expect(d).not.toBeNull();
-		expect(d!.year).toBe(1900);
-		expect(d!.month).toBe(3);
-		expect(d!.day).toBe(1);
-	});
-
-	it("out of range returns null", () => {
-		expect(parseExcelDateCode(3000000)).toBeNull();
-		expect(parseExcelDateCode(-1)).toBeNull();
-	});
-
-	it("1904 date system shifts by 1462 days", () => {
-		const d = parseExcelDateCode(1, { date1904: true });
-		expect(d).not.toBeNull();
-		// 1 + 1462 = 1463 serial in 1900 system = Jan 5, 1904
-		expect(d!.year).toBe(1904);
-	});
-
-	it("time components are extracted", () => {
-		// 0.5 = 12:00:00 (noon)
-		const d = parseExcelDateCode(0.5);
-		expect(d).not.toBeNull();
-		expect(d!.hours).toBe(12);
-		expect(d!.minutes).toBe(0);
-		expect(d!.seconds).toBe(0);
-	});
-
-	it("sub-second precision is preserved", () => {
-		// A value with fractional seconds
-		const d = parseExcelDateCode(44197.500001);
-		expect(d).not.toBeNull();
-		expect(d!.hours).toBe(12);
-	});
-});
-
-// ============================================================
-// src/ssf/format.ts — isDateFormat
-// ============================================================
-describe("SSF: isDateFormat", () => {
-	it("recognizes yyyy-mm-dd", () => {
-		expect(isDateFormat("yyyy-mm-dd")).toBe(true);
-	});
-
-	it("recognizes hh:mm:ss", () => {
-		expect(isDateFormat("hh:mm:ss")).toBe(true);
-	});
-
-	it("recognizes m/d/yy", () => {
-		expect(isDateFormat("m/d/yy")).toBe(true);
-	});
-
-	it("recognizes [h]:mm:ss", () => {
-		expect(isDateFormat("[h]:mm:ss")).toBe(true);
-	});
-
-	it("rejects #,##0.00", () => {
-		expect(isDateFormat("#,##0.00")).toBe(false);
-	});
-
-	it("rejects 0.00%", () => {
-		expect(isDateFormat("0.00%")).toBe(false);
-	});
-
-	it("rejects plain @", () => {
-		expect(isDateFormat("@")).toBe(false);
-	});
-
-	it("recognizes General as not date", () => {
-		expect(isDateFormat("General")).toBe(false);
-	});
-
-	it('skips quoted strings: "yy" is not date', () => {
-		expect(isDateFormat('"yy"')).toBe(false);
-	});
-
-	it("recognizes B1/B2 calendar modifier as date", () => {
-		expect(isDateFormat("B1yyyy-mm-dd")).toBe(true);
-		expect(isDateFormat("B2yyyy-mm-dd")).toBe(true);
-	});
-
-	it("recognizes AM/PM as date", () => {
-		expect(isDateFormat("h AM/PM")).toBe(true);
-	});
-
-	it("handles escaped characters", () => {
-		expect(isDateFormat("\\d")).toBe(false);
-	});
-
-	it("handles numeric characters", () => {
-		expect(isDateFormat("123")).toBe(false);
-	});
-
-	it("handles ? placeholder", () => {
-		expect(isDateFormat("# ?/?")).toBe(false);
-	});
-
-	it("handles * fill character", () => {
-		expect(isDateFormat("* #,##0")).toBe(false);
-	});
-
-	it("handles parentheses", () => {
-		expect(isDateFormat("(#,##0)")).toBe(false);
-	});
-});
-
-// ============================================================
-// src/ssf/format.ts — Number formatting
-// ============================================================
-describe("SSF: number format patterns", () => {
-	it("#,##0 — thousands separator exercises comma path", () => {
-		const result = formatNumber("#,##0", 1234567);
-		expect(result).toContain(",");
-	});
-
-	it("#,##0 — zero", () => {
-		const result = formatNumber("#,##0", 0);
-		expect(result).toContain("0");
-	});
-
-	it("#,##0.00 — two decimals", () => {
-		expect(formatNumber("#,##0.00", 1234.5)).toBe("1,234.50");
-	});
-
-	it("#,##0.00 — negative", () => {
-		expect(formatNumber("#,##0.00", -1234.5)).toBe("-1,234.50");
-	});
-
-	it("0.00E+00 — scientific", () => {
-		const result = formatNumber("0.00E+00", 12345);
-		expect(result).toMatch(/1\.23E\+04/i);
-	});
-
-	it("0.00E+00 — small number", () => {
-		const result = formatNumber("0.00E+00", 0.0012345);
-		expect(result).toMatch(/1\.23E-03/i);
-	});
-
-	it("##0.0E+0 — engineering notation", () => {
-		const result = formatNumber("##0.0E+0", 12345);
-		expect(result).toMatch(/E/);
-	});
-
-	it("0.00% — percentage", () => {
-		expect(formatNumber("0.00%", 0.1234)).toBe("12.34%");
-	});
-
-	it("0% — percentage exercises percent path", () => {
-		const result = formatNumber("0%", 0.5);
-		expect(result).toMatch(/\d/);
-	});
-
-	it("# ?/? — simple fraction", () => {
-		const result = formatNumber("# ?/?", 3.5);
-		expect(result).toContain("1/2");
-	});
-
-	it("# ??/?? — two-digit fraction", () => {
-		const result = formatNumber("# ??/??", 3.333);
-		expect(result).toContain("/");
-	});
-
-	it("# ?/8 — fixed denominator", () => {
-		const result = formatNumber("# ?/8", 3.25);
-		expect(result).toContain("/8");
-	});
-
-	it("$#,##0 — dollar", () => {
-		expect(formatNumber("$#,##0", 1234)).toBe("$1,234");
-	});
-
-	it("$#,##0.00 — dollar with decimals", () => {
-		expect(formatNumber("$#,##0.00", 1234.5)).toBe("$1,234.50");
-	});
-
-	it("#, — trailing comma exercises scaling path", () => {
-		const result = formatNumber("#,", 1234567);
-		expect(result).toMatch(/\d/);
-	});
-
-	it("#,, — trailing two commas exercises double scaling", () => {
-		const result = formatNumber("#,,", 1234567890);
-		expect(result).toMatch(/\d/);
-	});
-
-	it("000-00-0000 — dash-separated (SSN)", () => {
-		expect(formatNumber("000-00-0000", 123456789)).toBe("123-45-6789");
-	});
-
-	it("00 — zero-padded integer", () => {
-		expect(formatNumber("00", 5)).toBe("05");
-		expect(formatNumber("00", 99)).toBe("99");
-	});
-
-	it("0.00 — fixed two decimals", () => {
-		expect(formatNumber("0.00", 3.1)).toBe("3.10");
-		expect(formatNumber("0.00", 3)).toBe("3.00");
-	});
-
-	it("?? — space-padded", () => {
-		const result = formatNumber("??", 5);
-		expect(result).toBe(" 5");
-	});
-
-	it("General for integers", () => {
-		expect(formatNumber("General", 42)).toBe("42");
-	});
-
-	it("General for float", () => {
-		expect(formatNumber("General", 3.14)).toBe("3.14");
-	});
-
-	it("General for large number", () => {
-		const result = formatNumber("General", 1e15);
-		expect(result).toContain("E");
-	});
-
-	it("General for small number", () => {
-		const result = formatNumber("General", 0.00001);
-		expect(result).toBe("0.00001");
-	});
-
-	it("General for NaN", () => {
-		expect(formatNumber("General", NaN)).toBe("#NUM!");
-	});
-
-	it("General for Infinity", () => {
-		expect(formatNumber("General", Infinity)).toBe("#DIV/0!");
-	});
-
-	it("General for boolean true", () => {
-		expect(formatNumber("General", true)).toBe("TRUE");
-	});
-
-	it("General for boolean false", () => {
-		expect(formatNumber("General", false)).toBe("FALSE");
-	});
-
-	it("General for string", () => {
-		expect(formatNumber("General", "hello")).toBe("hello");
-	});
-
-	it("General for null", () => {
-		expect(formatNumber("General", null)).toBe("");
-	});
-
-	it("General for undefined", () => {
-		expect(formatNumber("General", undefined)).toBe("");
-	});
-
-	it("General for Date object", () => {
-		const result = formatNumber("General", new Date(2021, 0, 1));
-		expect(result).toMatch(/\d/);
-	});
-
-	it("text format @ passes through", () => {
-		expect(formatNumber("@", "hello")).toBe("hello");
-	});
-
-	it("format index lookup", () => {
-		// Format index 1 = "0"
-		const r1 = formatNumber(1, 42);
-		expect(r1).toMatch(/\d/);
-		// Format index 2 = "0.00"
-		const r2 = formatNumber(2, 3.14);
-		expect(r2).toContain("3");
-	});
-
-	it("format index 14 with dateNF override", () => {
-		const result = formatNumber(14, 44197, { dateNF: "yyyy-mm-dd" });
-		expect(result).toBe("2021-01-01");
-	});
-
-	it("format string m/d/yy with dateNF override", () => {
-		const result = formatNumber("m/d/yy", 44197, { dateNF: "yyyy-mm-dd" });
-		expect(result).toBe("2021-01-01");
-	});
-
-	it("empty/null value returns empty string", () => {
-		expect(formatNumber("0.00", "")).toBe("");
-		expect(formatNumber("0.00", null)).toBe("");
-	});
-
-	it("NaN with number format returns #NUM!", () => {
-		expect(formatNumber("0.00", NaN)).toBe("#NUM!");
-	});
-
-	it("Infinity with number format returns #DIV/0!", () => {
-		expect(formatNumber("0.00", Infinity)).toBe("#DIV/0!");
-	});
-
-	it("boolean true with format", () => {
-		expect(formatNumber("0.00", true)).toBe("TRUE");
-	});
-
-	it("boolean false with format", () => {
-		expect(formatNumber("0.00", false)).toBe("FALSE");
-	});
-
-	it("Date object with format", () => {
-		const result = formatNumber("yyyy-mm-dd", new Date(2021, 0, 1));
-		expect(result).toMatch(/2021/);
-	});
-
-	it("parenthesized negative format", () => {
-		const result = formatNumber("#,##0_);(#,##0)", -1234);
-		expect(result).toBe("(1,234)");
-	});
-
-	it("parenthesized positive format", () => {
-		const result = formatNumber("#,##0_);(#,##0)", 1234);
-		expect(result).toBe("1,234 ");
-	});
-
-	it("##,###  special case", () => {
-		expect(formatNumber("##,###", 1234)).toBe("1,234");
-	});
-
-	it("#,### special case", () => {
-		expect(formatNumber("#,###", 1234)).toBe("1,234");
-	});
-
-	it("###,### special case", () => {
-		expect(formatNumber("###,###", 1234)).toBe("1,234");
-	});
-
-	it("###,###.00", () => {
-		expect(formatNumber("###,###.00", 1234.5)).toBe("1,234.50");
-	});
-
-	it("#,###.00", () => {
-		expect(formatNumber("#,###.00", 1234.5)).toBe("1,234.50");
-	});
-
-	it("###,##0.00", () => {
-		expect(formatNumber("###,##0.00", 1234.5)).toBe("1,234.50");
-	});
-
-	it("multi-section format exercises section selection", () => {
-		// Use format 0 (single digit) to exercise the section selection paths
-		const pos = formatNumber("0;0;0", 5);
-		expect(pos).toMatch(/\d/);
-		const neg = formatNumber("0;0;0", -5);
-		expect(neg).toMatch(/\d/);
-		const zero = formatNumber("0;0;0", 0);
-		expect(zero).toBe("0");
-	});
-
-	it("4-section format with text", () => {
-		const result = formatNumber('#,##0;-#,##0;0;"text: "@', "hello");
-		expect(result).toBe("text: hello");
-	});
-
-	it("conditional format exercises conditional paths", () => {
-		const big = formatNumber('[>=100]"big";[<0]"neg";"small"', 200);
-		expect(big).toContain("big");
-		// Negative with conditional — exercises chkcond path
-		const neg = formatNumber('[>=100]"big";[<0]"neg";"small"', -5);
-		// May or may not output "neg" depending on how negatives interact with conditionals
-		expect(typeof neg).toBe("string");
-		const small = formatNumber('[>=100]"big";[<0]"neg";"small"', 50);
-		expect(small).toContain("small");
-	});
-
-	it("currency with locale [$€-407]", () => {
-		const result = formatNumber("[$€-407]#,##0.00", 1234.5);
-		expect(result).toBe("€1,234.50");
-	});
-
-	it("escaped character in format", () => {
-		const result = formatNumber("0\\-0", 12);
-		expect(result).toContain("-");
-	});
-
-	it("quoted literal in format", () => {
-		const result = formatNumber('#0" items"', 42);
-		expect(result).toContain("items");
-	});
-
-	it("_ padding character", () => {
-		const result = formatNumber("#0_)", 42);
-		expect(result).toContain("42");
-	});
-
-	it("#0 (leading # with 0)", () => {
-		expect(formatNumber("#0", 5)).toBe("5");
-		expect(formatNumber("#0", 42)).toBe("42");
-	});
-
-	it("0. format (decimal with no fractional part)", () => {
-		expect(formatNumber("0.#", 3)).toBe("3.");
-	});
-
-	it("fraction without whole: ?/?", () => {
-		const result = formatNumber("?/?", 0.5);
-		expect(result).toContain("1/2");
-	});
-
-	it("fraction for integer shows blank fraction", () => {
-		const result = formatNumber("# ?/?", 5);
-		expect(result).toContain("5");
-	});
-
-	it("00,000.00 format", () => {
-		const result = formatNumber("00,000.00", 1234.5);
-		expect(result).toMatch(/01,234\.50/);
-	});
-
-	it("sub-second format ss.000", () => {
-		const result = formatNumber("ss.000", 44197.500115);
-		expect(result).toBeTruthy();
-	});
-
-	it("era year format e", () => {
-		const result = formatNumber("e", 44197);
-		expect(result).toMatch(/\d{4}/);
-	});
-});
-
-// ============================================================
-// src/ssf/format.ts — General format edge cases
-// ============================================================
-describe("SSF: General format edge cases", () => {
-	it("very small negative", () => {
-		const result = formatNumber("General", -0.001);
-		expect(result).toBe("-0.001");
-	});
-
-	it("boundary: 10^10", () => {
-		const result = formatNumber("General", 1e10);
-		expect(result).toBe("10000000000");
-	});
-
-	it("boundary: 10^11", () => {
-		const result = formatNumber("General", 1e11);
-		expect(result).toMatch(/\d/);
-	});
-
-	it("large negative", () => {
-		const result = formatNumber("General", -1e15);
-		expect(result).toContain("E");
-	});
-
-	it("integer pass-through", () => {
-		expect(formatNumber("General", 0)).toBe("0");
-		expect(formatNumber("General", -42)).toBe("-42");
-	});
-});
-
-// ============================================================
-// XLSX roundtrip — workbook features
-// ============================================================
-describe("XLSX roundtrip: workbook features", () => {
-	it("hidden sheets survive roundtrip", async () => {
-		const ws1 = arrayToSheet([["Visible"]]);
-		const ws2 = arrayToSheet([["Hidden"]]);
-		const wb = createWorkbook(ws1, "Vis");
-		appendSheet(wb, ws2, "Hid");
-		setSheetVisibility(wb, "Hid", 1);
-
-		const bytes = await write(wb);
-		const wb2 = await read(bytes);
-		expect(wb2.SheetNames).toContain("Vis");
-		expect(wb2.SheetNames).toContain("Hid");
-		// Check hidden state via Workbook metadata
-		if (wb2.Workbook?.Sheets) {
-			const hidSheet = wb2.Workbook.Sheets.find((s: any) => s.name === "Hid");
-			if (hidSheet) {
-				expect(hidSheet.Hidden).toBe(1);
-			}
-		}
-	});
-
-	it("very hidden sheets survive roundtrip", async () => {
-		const ws1 = arrayToSheet([["A"]]);
-		const ws2 = arrayToSheet([["B"]]);
-		const wb = createWorkbook(ws1, "S1");
-		appendSheet(wb, ws2, "S2");
-		setSheetVisibility(wb, "S2", 2);
-
-		const bytes = await write(wb);
-		const wb2 = await read(bytes);
-		expect(wb2.SheetNames).toContain("S2");
-	});
-
-	it("defined names survive roundtrip", async () => {
-		const ws = arrayToSheet([["A"]]);
-		const wb = createWorkbook(ws, "Sheet1");
-		wb.Workbook = {
-			WBProps: {},
-			Sheets: [],
-			Names: [
-				{ Name: "MyRange", Ref: "Sheet1!$A$1:$B$2" },
-				{ Name: "HiddenName", Ref: "Sheet1!$C$1", Hidden: true },
-				{ Name: "LocalName", Ref: "Sheet1!$D$1", Sheet: 0 },
-			],
-		};
-
-		const bytes = await write(wb);
-		const wb2 = await read(bytes);
-		expect(wb2.Workbook?.Names).toBeDefined();
-		expect(wb2.Workbook!.Names!.length).toBeGreaterThanOrEqual(3);
-		const myRange = wb2.Workbook!.Names!.find((n: any) => n.Name === "MyRange");
-		expect(myRange).toBeDefined();
-		expect(myRange!.Ref).toContain("Sheet1");
-	});
-
-	it("merge cells survive roundtrip", async () => {
-		const ws = arrayToSheet([
-			["Merged", null, null],
-			[null, null, null],
-			["Normal", "B", "C"],
-		]);
-		ws["!merges"] = [{ s: { r: 0, c: 0 }, e: { r: 1, c: 2 } }];
-		const wb = createWorkbook(ws, "Sheet1");
-
-		const bytes = await write(wb);
-		const wb2 = await read(bytes);
-		const ws2 = wb2.Sheets[wb2.SheetNames[0]];
-		expect(ws2["!merges"]).toBeDefined();
-		expect(ws2["!merges"]!.length).toBe(1);
-		expect(ws2["!merges"]![0].s.r).toBe(0);
-		expect(ws2["!merges"]![0].e.r).toBe(1);
-	});
-
-	it("column widths survive roundtrip", async () => {
-		const ws = arrayToSheet([["A", "B", "C"]]);
-		ws["!cols"] = [{ width: 20 }, { width: 30 }, { width: 10 }];
-		const wb = createWorkbook(ws, "Sheet1");
-
-		const bytes = await write(wb);
-		const wb2 = await read(bytes, { cellStyles: true });
-		const ws2 = wb2.Sheets[wb2.SheetNames[0]];
-		expect(ws2["!cols"]).toBeDefined();
-		expect(ws2["!cols"]!.length).toBeGreaterThan(0);
-	});
-
-	it("row heights survive roundtrip", async () => {
-		const ws = arrayToSheet([["A"], ["B"], ["C"]]);
-		ws["!rows"] = [{ hpt: 30 }, undefined as any, { hpt: 40 }];
-		const wb = createWorkbook(ws, "Sheet1");
-
-		const bytes = await write(wb);
-		const wb2 = await read(bytes);
-		const ws2 = wb2.Sheets[wb2.SheetNames[0]];
-		expect(ws2["!rows"]).toBeDefined();
-		expect(ws2["!rows"]![0].hpt).toBe(30);
-	});
-
-	it("hidden rows survive roundtrip", async () => {
-		const ws = arrayToSheet([["Visible"], ["Hidden"], ["Visible2"]]);
-		ws["!rows"] = [undefined as any, { hidden: true }, undefined as any];
-		const wb = createWorkbook(ws, "Sheet1");
-
-		const bytes = await write(wb);
-		const wb2 = await read(bytes);
-		const ws2 = wb2.Sheets[wb2.SheetNames[0]];
-		expect(ws2["!rows"]).toBeDefined();
-		expect(ws2["!rows"]![1].hidden).toBe(true);
-	});
-
-	it("autofilter survives roundtrip", async () => {
-		const ws = arrayToSheet([
-			["Name", "Age"],
-			["Alice", 30],
-			["Bob", 25],
-		]);
-		ws["!autofilter"] = { ref: "A1:B3" };
-		const wb = createWorkbook(ws, "Sheet1");
-
-		const bytes = await write(wb);
-		const wb2 = await read(bytes);
-		const ws2 = wb2.Sheets[wb2.SheetNames[0]];
-		expect(ws2["!autofilter"]).toBeDefined();
-		expect(ws2["!autofilter"]!.ref).toBe("A1:B3");
-	});
-
-	it("page margins survive roundtrip", async () => {
-		const ws = arrayToSheet([["A"]]);
-		ws["!margins"] = { left: 1, right: 1, top: 1.5, bottom: 1.5, header: 0.5, footer: 0.5 };
-		const wb = createWorkbook(ws, "Sheet1");
-
-		const bytes = await write(wb);
-		const wb2 = await read(bytes);
-		const ws2 = wb2.Sheets[wb2.SheetNames[0]];
-		expect(ws2["!margins"]).toBeDefined();
-		expect(ws2["!margins"]!.left).toBe(1);
-	});
-
-	it("formulas survive roundtrip", async () => {
-		const ws: any = {
-			A1: { t: "n", v: 1 },
-			A2: { t: "n", v: 2 },
-			A3: { t: "n", v: 3, f: "SUM(A1:A2)" },
-			"!ref": "A1:A3",
-		};
-		const wb = createWorkbook(ws, "Sheet1");
-
-		const bytes = await write(wb);
-		const wb2 = await read(bytes, { cellFormula: true });
-		const ws2 = wb2.Sheets[wb2.SheetNames[0]];
-		const a3 = (ws2 as any)["A3"];
-		expect(a3).toBeDefined();
-		expect(a3.f).toBe("SUM(A1:A2)");
-	});
-
-	it("array formulas survive roundtrip", async () => {
-		const ws = arrayToSheet([
-			[1, 10],
-			[2, 20],
-			[3, 30],
-		]);
-		setArrayFormula(ws, "C1:C3", "A1:A3*B1:B3");
-
-		const wb = createWorkbook(ws, "Sheet1");
-		const bytes = await write(wb);
-		const wb2 = await read(bytes, { cellFormula: true });
-		const ws2 = wb2.Sheets[wb2.SheetNames[0]];
-		const c1 = (ws2 as any)["C1"];
-		expect(c1).toBeDefined();
-		expect(c1.f).toBe("A1:A3*B1:B3");
-		expect(c1.F).toBe("C1:C3");
-	});
-
-	it("boolean cells survive roundtrip", async () => {
-		const ws: any = { A1: { t: "b", v: true }, A2: { t: "b", v: false }, "!ref": "A1:A2" };
-		const wb = createWorkbook(ws, "Sheet1");
-		const bytes = await write(wb);
-		const wb2 = await read(bytes);
-		const ws2 = wb2.Sheets[wb2.SheetNames[0]];
-		expect((ws2 as any)["A1"].v).toBe(true);
-		expect((ws2 as any)["A2"].v).toBe(false);
-	});
-
-	it("error cells survive roundtrip", async () => {
-		const ws: any = { A1: { t: "e", v: 0x07, w: "#DIV/0!" }, "!ref": "A1:A1" };
-		const wb = createWorkbook(ws, "Sheet1");
-		const bytes = await write(wb);
-		const wb2 = await read(bytes);
-		const ws2 = wb2.Sheets[wb2.SheetNames[0]];
-		expect((ws2 as any)["A1"].t).toBe("e");
-	});
-
-	it("date cells with cellDates option", async () => {
-		const ws: any = { A1: { t: "d", v: new Date("2021-06-15T00:00:00") }, "!ref": "A1:A1" };
-		const wb = createWorkbook(ws, "Sheet1");
-		const bytes = await write(wb, { cellDates: true });
-		const wb2 = await read(bytes, { cellDates: true });
-		const ws2 = wb2.Sheets[wb2.SheetNames[0]];
-		const a1 = (ws2 as any)["A1"];
-		expect(a1).toBeDefined();
-	});
-
-	it("multiple sheets", async () => {
-		const ws1 = arrayToSheet([["First"]]);
-		const ws2 = arrayToSheet([["Second"]]);
-		const ws3 = arrayToSheet([["Third"]]);
-		const wb = createWorkbook(ws1, "S1");
-		appendSheet(wb, ws2, "S2");
-		appendSheet(wb, ws3, "S3");
-
-		const bytes = await write(wb);
-		const wb2 = await read(bytes);
-		expect(wb2.SheetNames).toEqual(["S1", "S2", "S3"]);
-	});
-
-	it("inline strings survive roundtrip", async () => {
-		const ws = arrayToSheet([["Hello world", "Test 123"]]);
-		const wb = createWorkbook(ws, "Sheet1");
-
-		const bytes = await write(wb);
-		const wb2 = await read(bytes);
-		const rows = sheetToJson(wb2.Sheets[wb2.SheetNames[0]], { header: 1 });
-		expect(rows[0]).toContain("Hello world");
-		expect(rows[0]).toContain("Test 123");
-	});
-
-	it("bookSheets option returns only sheet names", async () => {
-		const ws = arrayToSheet([["A"]]);
-		const wb = createWorkbook(ws, "MySheet");
-		const bytes = await write(wb);
-		const wb2 = await read(bytes, { bookSheets: true });
-		expect(wb2.SheetNames).toContain("MySheet");
-		expect(Object.keys(wb2.Sheets || {})).toHaveLength(0);
-	});
-
-	it("bookProps option returns properties", async () => {
-		const ws = arrayToSheet([["A"]]);
-		const wb = createWorkbook(ws, "Sheet1");
-		wb.Props = { Title: "Test Title", Author: "Test Author" };
-		const bytes = await write(wb);
-		const wb2 = await read(bytes, { bookProps: true });
-		expect(wb2.Props).toBeDefined();
-	});
-
-	it("sheetRows limits row count", async () => {
-		const data = Array.from({ length: 50 }, (_, i) => [i]);
-		const ws = arrayToSheet(data);
-		const wb = createWorkbook(ws, "Sheet1");
-		const bytes = await write(wb);
-		const wb2 = await read(bytes, { sheetRows: 10 });
-		const ws2 = wb2.Sheets[wb2.SheetNames[0]];
-		const rows = sheetToJson(ws2, { header: 1 });
-		expect(rows.length).toBeLessThanOrEqual(10);
-	});
-
-	it("sheets filter by name", async () => {
-		const ws1 = arrayToSheet([["First"]]);
-		const ws2 = arrayToSheet([["Second"]]);
-		const wb = createWorkbook(ws1, "S1");
-		appendSheet(wb, ws2, "S2");
-		const bytes = await write(wb);
-		const wb2 = await read(bytes, { sheets: "S2" } as any);
-		expect(wb2.SheetNames).toContain("S2");
-		// S1 might still be in SheetNames (from workbook.xml) but its data shouldn't be loaded
-		if (wb2.SheetNames.includes("S1")) {
-			expect(wb2.Sheets["S1"]).toBeUndefined();
-		}
-	});
-
-	it("sheets filter by index", async () => {
-		const ws1 = arrayToSheet([["First"]]);
-		const ws2 = arrayToSheet([["Second"]]);
-		const wb = createWorkbook(ws1, "S1");
-		appendSheet(wb, ws2, "S2");
-		const bytes = await write(wb);
-		const wb2 = await read(bytes, { sheets: 1 } as any);
-		expect(wb2.Sheets["S2"]).toBeDefined();
-	});
-
-	it("sheets filter by array", async () => {
-		const ws1 = arrayToSheet([["First"]]);
-		const ws2 = arrayToSheet([["Second"]]);
-		const ws3 = arrayToSheet([["Third"]]);
-		const wb = createWorkbook(ws1, "S1");
-		appendSheet(wb, ws2, "S2");
-		appendSheet(wb, ws3, "S3");
-		const bytes = await write(wb);
-		const wb2 = await read(bytes, { sheets: [0, "S3"] } as any);
-		expect(wb2.Sheets["S1"]).toBeDefined();
-		expect(wb2.Sheets["S3"]).toBeDefined();
-	});
-
-	it("custom properties survive roundtrip", async () => {
-		const ws = arrayToSheet([["A"]]);
-		const wb = createWorkbook(ws, "Sheet1");
-		wb.Custprops = { myKey: "myValue", myNumber: 42 };
-		const bytes = await write(wb);
-		const wb2 = await read(bytes);
-		expect(wb2.Custprops).toBeDefined();
-		expect(wb2.Custprops!.myKey).toBe("myValue");
-	});
-});
-
-// ============================================================
-// src/xlsx/workbook.ts — direct tests
-// ============================================================
-describe("workbook.ts: validation", () => {
-	it("validateSheetName rejects empty name", () => {
-		expect(() => validateSheetName("")).toThrow("blank");
-	});
-
-	it("validateSheetName rejects too-long name", () => {
-		expect(() => validateSheetName("a".repeat(32))).toThrow("exceed 31");
-	});
-
-	it("validateSheetName rejects apostrophe boundaries", () => {
-		expect(() => validateSheetName("'Name")).toThrow("apostrophe");
-		expect(() => validateSheetName("Name'")).toThrow("apostrophe");
-	});
-
-	it("validateSheetName rejects 'History'", () => {
-		expect(() => validateSheetName("history")).toThrow("History");
-	});
-
-	it("validateSheetName rejects forbidden characters", () => {
-		expect(() => validateSheetName("Sheet:1")).toThrow();
-		expect(() => validateSheetName("Sheet[1]")).toThrow();
-		expect(() => validateSheetName("Sheet*")).toThrow();
-		expect(() => validateSheetName("Sheet?")).toThrow();
-		expect(() => validateSheetName("Sheet/1")).toThrow();
-		expect(() => validateSheetName("Sheet\\1")).toThrow();
-	});
-
-	it("validateSheetName safe mode returns false", () => {
-		expect(validateSheetName("", true)).toBe(false);
-		expect(validateSheetName("a".repeat(32), true)).toBe(false);
-	});
-
-	it("validateWorkbook rejects invalid structure", () => {
-		expect(() => {
-			validateWorkbook(null as any);
-		}).toThrow("Invalid");
-		expect(() => {
-			validateWorkbook({ SheetNames: [], Sheets: {} });
-		}).toThrow("empty");
-		expect(() => {
-			validateWorkbook({ SheetNames: ["A", "A"], Sheets: {} });
-		}).toThrow("Duplicate");
-	});
-
-	it("is1904DateSystem returns false for non-1904 workbooks", () => {
-		expect(is1904DateSystem({} as any)).toBe("false");
-		expect(is1904DateSystem({ Workbook: {} } as any)).toBe("false");
-		expect(is1904DateSystem({ Workbook: { WBProps: { date1904: false } } } as any)).toBe("false");
-	});
-
-	it("is1904DateSystem returns true for 1904 workbooks", () => {
-		expect(is1904DateSystem({ Workbook: { WBProps: { date1904: true } } } as any)).toBe("true");
-	});
-});
-
-// ============================================================
-// src/xlsx/worksheet.ts — parseWorksheetXml
-// ============================================================
-describe("worksheet.ts: parsing edge cases", () => {
-	it("empty data returns empty sheet", () => {
-		const ws = parseWorksheetXml("");
-		expect(ws).toBeDefined();
-	});
-
-	it("dense mode", () => {
-		const xml = `<?xml version="1.0"?>
-		<worksheet xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main">
-		<sheetData>
-		<row r="1"><c r="A1" t="str"><v>hello</v></c></row>
-		</sheetData>
-		</worksheet>`;
-		const ws = parseWorksheetXml(xml, { dense: true });
-		expect(ws["!data"]).toBeDefined();
-		expect(ws["!data"]![0][0].v).toBe("hello");
-	});
-
-	it("cell types: boolean, error, inline string", () => {
-		const xml = `<?xml version="1.0"?>
-		<worksheet xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main">
-		<sheetData>
-		<row r="1">
-			<c r="A1" t="b"><v>1</v></c>
-			<c r="B1" t="e"><v>#DIV/0!</v></c>
-			<c r="C1" t="inlineStr"><is><t>Inline</t></is></c>
-		</row>
-		</sheetData>
-		</worksheet>`;
-		const ws = parseWorksheetXml(xml);
-		expect((ws as any)["A1"].t).toBe("b");
-		expect((ws as any)["A1"].v).toBe(true);
-		expect((ws as any)["B1"].t).toBe("e");
-		expect((ws as any)["C1"].t).toBe("s");
-		expect((ws as any)["C1"].v).toBe("Inline");
-	});
-
-	it("sheetRows limits parsing", () => {
-		const xml = `<?xml version="1.0"?>
-		<worksheet xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main">
-		<dimension ref="A1:A10"/>
-		<sheetData>
-		<row r="1"><c r="A1"><v>1</v></c></row>
-		<row r="2"><c r="A2"><v>2</v></c></row>
-		<row r="3"><c r="A3"><v>3</v></c></row>
-		<row r="10"><c r="A10"><v>10</v></c></row>
-		</sheetData>
-		</worksheet>`;
-		const ws = parseWorksheetXml(xml, { sheetRows: 2 });
-		expect((ws as any)["A1"]).toBeDefined();
-		expect((ws as any)["A3"]).toBeUndefined();
-	});
-
-	it("sheetStubs creates z-type cells", () => {
-		const xml = `<?xml version="1.0"?>
-		<worksheet xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main">
-		<sheetData>
-		<row r="1"><c r="A1"></c></row>
-		</sheetData>
-		</worksheet>`;
-		const ws = parseWorksheetXml(xml, { sheetStubs: true });
-		expect((ws as any)["A1"]).toBeDefined();
-		expect((ws as any)["A1"].t).toBe("z");
-	});
-
-	it("resolveSharedStrings works in sparse mode", () => {
-		const ws: any = {
-			A1: { t: "s", v: "", _sstIdx: 0 },
-			B1: { t: "n", v: 42 },
-			"!ref": "A1:B1",
-		};
-		const sst: any = [{ t: "Hello", h: "<b>Hello</b>", r: "<t>Hello</t>" }];
-		resolveSharedStrings(ws, sst, {});
-		expect(ws.A1.v).toBe("Hello");
-	});
-
-	it("resolveSharedStrings works in dense mode", () => {
-		const ws: any = {
-			"!data": [
-				[
-					{ t: "s", v: "", _sstIdx: 0 },
-					{ t: "n", v: 42 },
-				],
-			],
-			"!ref": "A1:B1",
-		};
-		const sst: any = [{ t: "World", h: "World", r: "<t>World</t>" }];
-		resolveSharedStrings(ws, sst, {});
-		expect(ws["!data"][0][0].v).toBe("World");
-	});
-});
-
-// ============================================================
-// src/xlsx/shared-strings.ts — write/parse SST
-// ============================================================
-describe("shared-strings.ts: write and parse", () => {
-	it("writeSstXml with bookSST=false returns empty", () => {
-		expect(writeSstXml([] as any, { bookSST: false })).toBe("");
-	});
-
-	it("roundtrip plain text SST", () => {
-		const sst: any = [{ t: "Hello" }, { t: "World" }];
-		sst.Count = 2;
-		sst.Unique = 2;
-		const xml = writeSstXml(sst, { bookSST: true });
-		expect(xml).toContain("<sst");
-		expect(xml).toContain("Hello");
-		expect(xml).toContain("World");
-
-		const parsed = parseSstXml(xml);
-		expect(parsed[0].t).toBe("Hello");
-		expect(parsed[1].t).toBe("World");
-	});
-
-	it("roundtrip SST with whitespace-preserving text", () => {
-		const sst: any = [{ t: "  leading spaces" }, { t: "trailing  " }];
-		sst.Count = 2;
-		sst.Unique = 2;
-		const xml = writeSstXml(sst, { bookSST: true });
-		expect(xml).toContain('xml:space="preserve"');
-
-		const parsed = parseSstXml(xml);
-		expect(parsed[0].t).toBe("  leading spaces");
-	});
-
-	it("handles null entries", () => {
-		const sst: any = [{ t: "A" }, null, { t: "C" }];
-		sst.Count = 3;
-		sst.Unique = 2;
-		const xml = writeSstXml(sst, { bookSST: true });
-		expect(xml).toContain("A");
-		expect(xml).toContain("C");
-	});
-
-	it("empty data returns empty array", () => {
-		const parsed = parseSstXml("");
-		expect(parsed).toHaveLength(0);
-	});
-});
-
-// ============================================================
-// src/api/csv.ts — edge cases
-// ============================================================
-describe("csv.ts: advanced CSV features", () => {
-	it("sheetToCsv with skipHidden rows", () => {
-		const ws = arrayToSheet([["A"], ["B"], ["C"]]);
-		ws["!rows"] = [undefined as any, { hidden: true }, undefined as any];
-		const csv = sheetToCsv(ws, { skipHidden: true });
-		expect(csv).toContain("A");
-		expect(csv).not.toContain("B");
-		expect(csv).toContain("C");
-	});
-
-	it("sheetToCsv with skipHidden cols", () => {
-		const ws = arrayToSheet([["A", "B", "C"]]);
-		ws["!cols"] = [undefined as any, { hidden: true }, undefined as any];
-		const csv = sheetToCsv(ws, { skipHidden: true });
-		expect(csv).toContain("A");
-		expect(csv).not.toContain("B");
-		expect(csv).toContain("C");
-	});
-
-	it("sheetToCsv with strip option", () => {
-		const ws = arrayToSheet([["A", "", ""]]);
-		const csv = sheetToCsv(ws, { strip: true });
-		expect(csv).toBe("A");
-	});
-
-	it("sheetToCsv with blankrows=false", () => {
-		const ws = arrayToSheet([["A"], [], ["C"]]);
-		const csv = sheetToCsv(ws, { blankrows: false });
-		expect(csv).toBe("A\nC");
-	});
-
-	it("sheetToCsv with forceQuotes", () => {
-		const ws = arrayToSheet([["Simple"]]);
-		const csv = sheetToCsv(ws, { forceQuotes: true });
-		expect(csv).toBe('"Simple"');
-	});
-
-	it("sheetToCsv with custom RS", () => {
-		const ws = arrayToSheet([["A"], ["B"]]);
-		const csv = sheetToCsv(ws, { RS: "\r\n" });
-		expect(csv).toBe("A\r\nB");
-	});
-
-	it("sheetToCsv with rawNumbers", () => {
-		const ws = arrayToSheet([[1.23456789]]);
-		const csv = sheetToCsv(ws, { rawNumbers: true });
-		expect(csv).toBe("1.23456789");
-	});
-
-	it("sheetToCsv quotes commas", () => {
-		const ws = arrayToSheet([["A,B"]]);
-		const csv = sheetToCsv(ws);
-		expect(csv).toBe('"A,B"');
-	});
-
-	it("sheetToCsv quotes newlines", () => {
-		const ws = arrayToSheet([["A\nB"]]);
-		const csv = sheetToCsv(ws);
-		expect(csv).toBe('"A\nB"');
-	});
-
-	it("sheetToCsv quotes double-quotes", () => {
-		const ws = arrayToSheet([['A"B']]);
-		const csv = sheetToCsv(ws);
-		expect(csv).toBe('"A""B"');
-	});
-
-	it("sheetToCsv quotes bare ID", () => {
-		const ws = arrayToSheet([["ID", "Name"]]);
-		const csv = sheetToCsv(ws);
-		expect(csv).toContain('"ID"');
-	});
-
-	it("sheetToCsv with formula-only cell", () => {
-		const ws: any = { A1: { t: "z", f: "SUM(B1:B10)" }, "!ref": "A1:A1" };
-		const csv = sheetToCsv(ws);
-		expect(csv).toContain("=SUM(B1:B10)");
-	});
-
-	it("sheetToCsv formula with comma gets quoted", () => {
-		const ws: any = { A1: { t: "z", f: "IF(A2,B2,C2)" }, "!ref": "A1:A1" };
-		const csv = sheetToCsv(ws);
-		expect(csv).toContain('"=IF(A2,B2,C2)"');
-	});
-
-	it("sheetToTxt produces tab-separated output", () => {
-		const ws = arrayToSheet([
-			["A", "B"],
-			[1, 2],
-		]);
-		const tsv = sheetToTxt(ws);
-		expect(tsv).toContain("A\tB");
-		expect(tsv).toContain("1\t2");
-	});
-
-	it("csvToSheet with tab separator", () => {
-		const ws = csvToSheet("A\tB\n1\t2", { FS: "\t" });
-		const rows = sheetToJson(ws, { header: 1 });
-		expect(rows[0]).toContain("A");
-		expect(rows[0]).toContain("B");
-	});
-
-	it("csvToSheet with CRLF line endings", () => {
-		const ws = csvToSheet("A,B\r\n1,2\r\n3,4");
-		const rows = sheetToJson(ws, { header: 1 });
-		expect(rows.length).toBeGreaterThanOrEqual(3);
-	});
-
-	it("csvToSheet with quoted fields containing newlines", () => {
-		const ws = csvToSheet('"Line1\nLine2",B\n1,2');
-		const rows = sheetToJson(ws, { header: 1 });
-		expect(rows[0][0]).toBe("Line1\nLine2");
-	});
-
-	it("csvToSheet with escaped double-quotes", () => {
-		const ws = csvToSheet('"He said ""hi""",B\n1,2');
-		const rows = sheetToJson(ws, { header: 1 });
-		expect(rows[0][0]).toBe('He said "hi"');
-	});
-
-	it("csvToSheet coerces booleans", () => {
-		const ws = csvToSheet("TRUE,FALSE,true,false");
-		const rows = sheetToJson(ws, { header: 1 });
-		expect(rows[0][0]).toBe(true);
-		expect(rows[0][1]).toBe(false);
-		expect(rows[0][2]).toBe(true);
-		expect(rows[0][3]).toBe(false);
-	});
-
-	it("null sheet returns empty string", () => {
-		expect(sheetToCsv(null as any)).toBe("");
-		expect(sheetToCsv({} as any)).toBe("");
-	});
-});
-
-// ============================================================
-// src/api/book.ts — uncovered branches
-// ============================================================
-describe("book.ts: API edge cases", () => {
-	it("appendSheet auto-generates name", () => {
-		const wb = createWorkbook();
-		appendSheet(wb, arrayToSheet([["A"]]));
-		expect(wb.SheetNames).toContain("Sheet1");
-		appendSheet(wb, arrayToSheet([["B"]]));
-		expect(wb.SheetNames).toContain("Sheet2");
-	});
-
-	it("appendSheet with roll option", () => {
-		const wb = createWorkbook(arrayToSheet([["A"]]), "Sheet1");
-		const name = appendSheet(wb, arrayToSheet([["B"]]), "Sheet1", true);
-		expect(name).toBe("Sheet2");
-		expect(wb.SheetNames).toContain("Sheet2");
-	});
-
-	it("appendSheet rejects duplicate without roll", () => {
-		const wb = createWorkbook(arrayToSheet([["A"]]), "Sheet1");
-		expect(() => appendSheet(wb, arrayToSheet([["B"]]), "Sheet1")).toThrow("already exists");
-	});
-
-	it("getSheetIndex finds by name", () => {
-		const wb = createWorkbook(arrayToSheet([["A"]]), "MySheet");
-		expect(getSheetIndex(wb, "MySheet")).toBe(0);
-	});
-
-	it("getSheetIndex finds by index", () => {
-		const wb = createWorkbook(arrayToSheet([["A"]]), "Sheet1");
-		expect(getSheetIndex(wb, 0)).toBe(0);
-	});
-
-	it("getSheetIndex throws for missing name", () => {
-		const wb = createWorkbook(arrayToSheet([["A"]]), "Sheet1");
-		expect(() => getSheetIndex(wb, "NoSuchSheet")).toThrow("Cannot find");
-	});
-
-	it("getSheetIndex throws for out-of-range index", () => {
-		const wb = createWorkbook(arrayToSheet([["A"]]), "Sheet1");
-		expect(() => getSheetIndex(wb, 99)).toThrow("Cannot find");
-	});
-
-	it("setSheetVisibility rejects invalid value", () => {
-		const wb = createWorkbook(arrayToSheet([["A"]]), "Sheet1");
-		expect(() => {
-			setSheetVisibility(wb, 0, 99 as any);
-		}).toThrow("Bad sheet visibility");
-	});
-
-	it("setSheetVisibility initializes Workbook metadata", () => {
-		const wb = createWorkbook(arrayToSheet([["A"]]), "Sheet1");
-		setSheetVisibility(wb, 0, 1);
-		expect(wb.Workbook).toBeDefined();
-		expect(wb.Workbook!.Sheets![0].Hidden).toBe(1);
-	});
-
-	it("createSheet creates empty sheet", () => {
-		const ws = createSheet();
-		expect(ws).toBeDefined();
-	});
-
-	it("createSheet dense mode", () => {
-		const ws = createSheet({ dense: true });
-		expect(ws["!data"]).toBeDefined();
-	});
-
-	it("setCellNumberFormat sets z property", () => {
-		const cell: any = { t: "n", v: 42 };
-		setCellNumberFormat(cell, "#,##0.00");
-		expect(cell.z).toBe("#,##0.00");
-	});
-
-	it("setCellHyperlink sets link", () => {
-		const cell: any = { t: "s", v: "Click" };
-		setCellHyperlink(cell, "https://example.com", "Example");
-		expect(cell.l.Target).toBe("https://example.com");
-		expect(cell.l.Tooltip).toBe("Example");
-	});
-
-	it("setCellHyperlink removes link when falsy", () => {
-		const cell: any = { t: "s", v: "Click", l: { Target: "https://example.com" } };
-		setCellHyperlink(cell);
-		expect(cell.l).toBeUndefined();
-	});
-
-	it("setCellInternalLink adds # prefix", () => {
-		const cell: any = { t: "s", v: "Click" };
-		setCellInternalLink(cell, "Sheet2!A1");
-		expect(cell.l.Target).toBe("#Sheet2!A1");
-	});
-
-	it("addCellComment adds comment", () => {
-		const cell: any = { t: "s", v: "Value" };
-		addCellComment(cell, "Check this", "Alice");
-		expect(cell.c).toHaveLength(1);
-		expect(cell.c[0].t).toBe("Check this");
-		expect(cell.c[0].a).toBe("Alice");
-	});
-
-	it("addCellComment uses default author", () => {
-		const cell: any = { t: "s", v: "Value" };
-		addCellComment(cell, "Note");
-		expect(cell.c[0].a).toBe("SheetJS");
-	});
-
-	it("sheetToFormulae extracts formulas", () => {
-		const ws: any = {
-			A1: { t: "n", v: 42 },
-			A2: { t: "s", v: "Hello" },
-			A3: { t: "b", v: true },
-			A4: { t: "n", v: 10, f: "SUM(A1:A1)" },
-			"!ref": "A1:A4",
-		};
-		const formulae = sheetToFormulae(ws);
-		expect(formulae.length).toBeGreaterThanOrEqual(4);
-		expect(formulae).toContain("A1=42");
-		expect(formulae).toContain("A2='Hello");
-		expect(formulae).toContain("A3=TRUE");
-		expect(formulae).toContain("A4=SUM(A1:A1)");
-	});
-
-	it("sheetToFormulae handles array formulas", () => {
-		const ws = arrayToSheet([[1], [2], [3]]);
-		setArrayFormula(ws, "B1:B3", "A1:A3*2");
-		const formulae = sheetToFormulae(ws);
-		const arrayFmla = formulae.find((f) => f.includes("A1:A3*2"));
-		expect(arrayFmla).toBeDefined();
-		expect(arrayFmla).toContain("B1:B3");
-	});
-
-	it("sheetToFormulae with dense mode", () => {
-		const ws = arrayToSheet([["A", 1]], { dense: true } as any);
-		const formulae = sheetToFormulae(ws);
-		expect(formulae.length).toBeGreaterThanOrEqual(2);
-	});
-
-	it("sheetToFormulae returns empty for null sheet", () => {
-		expect(sheetToFormulae(null as any)).toEqual([]);
-		expect(sheetToFormulae({} as any)).toEqual([]);
-	});
-
-	it("sheetToFormulae with w property", () => {
-		const ws: any = {
-			A1: { t: "n", v: 42, w: "42.00" },
-			"!ref": "A1:A1",
-		};
-		// When v is present, it uses v for numeric
-		const formulae = sheetToFormulae(ws);
-		expect(formulae).toContain("A1=42");
-	});
-
-	it("setArrayFormula dense mode", () => {
-		const ws = createSheet({ dense: true });
-		ws["!ref"] = "A1:A1";
-		(ws as any)["!data"] = [[{ t: "n", v: 1 }]];
-		setArrayFormula(ws, "B1:B2", "A1*2", true);
-		expect((ws as any)["!data"][0][1].f).toBe("A1*2");
-		expect((ws as any)["!data"][0][1].D).toBe(true);
-	});
-
-	it("setArrayFormula expands ref", () => {
-		const ws = arrayToSheet([[1]]);
-		setArrayFormula(ws, "C5:D6", "A1*2");
-		const ref = ws["!ref"];
-		expect(ref).toContain("D6");
-	});
-});
-
-// ============================================================
-// src/xlsx/write-zip.ts — writeZipXlsx additional paths
-// ============================================================
-describe("write-zip.ts: comments in roundtrip", () => {
-	it("legacy comments survive write → read", async () => {
-		const ws = arrayToSheet([["Value"]]);
-		// Prepare comments in the format expected by the writer
-		(ws as any)["!comments"] = [["A1", [{ a: "Alice", t: "A note", T: false }]]];
-		(ws as any)["!legacy"] = true;
-
-		const wb = createWorkbook(ws, "Sheet1");
-		const bytes = await write(wb);
-		const wb2 = await read(bytes);
-		const ws2 = wb2.Sheets[wb2.SheetNames[0]];
-		// Check that the comment was read back
-		const a1 = (ws2 as any)["A1"];
-		expect(a1).toBeDefined();
-		if (a1.c) {
-			expect(a1.c.length).toBeGreaterThanOrEqual(1);
-			expect(a1.c[0].t).toContain("A note");
-		}
-	});
-
-	it("threaded comments survive write → read", async () => {
-		const ws = arrayToSheet([["Value"]]);
-		(ws as any)["!comments"] = [
-			[
-				"A1",
-				[
-					{ a: "Alice", t: "First comment", T: true, ID: undefined },
-					{ a: "Bob", t: "Reply", T: true, ID: undefined },
-				],
-			],
-		];
-		(ws as any)["!legacy"] = true;
-
-		const wb = createWorkbook(ws, "Sheet1");
-		const bytes = await write(wb);
-		const wb2 = await read(bytes);
-		const ws2 = wb2.Sheets[wb2.SheetNames[0]];
-		const a1 = (ws2 as any)["A1"];
-		expect(a1).toBeDefined();
-		if (a1.c) {
-			expect(a1.c.length).toBeGreaterThanOrEqual(1);
-		}
-	});
-
-	it("workbook with veryHidden sheets filters app props", async () => {
-		const ws1 = arrayToSheet([["A"]]);
-		const ws2 = arrayToSheet([["B"]]);
-		const wb = createWorkbook(ws1, "Visible");
-		appendSheet(wb, ws2, "VeryHidden");
-		wb.Workbook = {
-			Sheets: [{ Hidden: 0 }, { Hidden: 2 }],
-		} as any;
-
-		const bytes = await write(wb);
-		const wb2 = await read(bytes);
-		expect(wb2.SheetNames).toContain("Visible");
-	});
-});
-
-// ============================================================
-// Additional SSF format.ts coverage for write_num_int paths
-// ============================================================
-describe("SSF: integer format paths", () => {
-	it("#,##0 for integer", () => {
-		expect(formatNumber("#,##0", 1234)).toBe("1,234");
-	});
-
-	it("#,##0 for negative integer", () => {
-		const result = formatNumber("#,##0", -1234);
-		expect(result).toContain("-");
-	});
-
-	it("#,##0.00 for integer", () => {
-		expect(formatNumber("#,##0.00", 1234)).toBe("1,234.00");
-	});
-
-	it("0.00 for integer", () => {
-		expect(formatNumber("0.00", 42)).toBe("42.00");
-	});
-
-	it("000 for small integer", () => {
-		expect(formatNumber("000", 5)).toBe("005");
-	});
-
-	it("# ?/16 for integer", () => {
-		const result = formatNumber("# ?/16", 5);
-		expect(result).toContain("5");
-	});
-
-	it("# ??/?? for integer", () => {
-		const result = formatNumber("# ??/??", 5);
-		expect(result).toContain("5");
-	});
-
-	it("?/? for integer", () => {
-		const result = formatNumber("?/?", 3);
-		expect(result).toContain("3");
-	});
-
-	it("#0 for integer", () => {
-		expect(formatNumber("#0", 42)).toBe("42");
-	});
-
-	it("0. for integer (trailing decimal)", () => {
-		const result = formatNumber("0.#", 42);
-		expect(result).toBe("42.");
-	});
-
-	it("00,000.00 for integer", () => {
-		const result = formatNumber("00,000.00", 1234);
-		expect(result).toMatch(/01,234\.00/);
-	});
-
-	it("#0.0# for integer", () => {
-		const result = formatNumber("#0.0#", 42);
-		expect(result).toMatch(/42\.0/);
-	});
-
-	it("0E+00 for integer", () => {
-		const result = formatNumber("0E+00", 12345);
-		expect(result).toMatch(/E/);
-	});
-
-	it("$0 for integer exercises dollar path", () => {
-		const result = formatNumber("$0", 42);
-		expect(result).toContain("$");
-	});
-
-	it("#, for integer exercises trailing comma path", () => {
-		const result = formatNumber("#,", 1234567);
-		expect(result).toMatch(/\d/);
-	});
-
-	it("0% for integer exercises percent path", () => {
-		const result = formatNumber("0%", 1);
-		expect(result).toMatch(/\d/);
-	});
-
-	it("(#,##0) parenthesized format for negative", () => {
-		const result = formatNumber("(#,##0)", -1234);
-		expect(result).toContain("(");
-	});
-
-	it("(#,##0) parenthesized format for positive", () => {
-		const result = formatNumber("(#,##0)", 1234);
-		expect(result).toMatch(/\d/);
-	});
-
-	it("000\\-00\\-0000 escaped dash format for integer", () => {
-		const result = formatNumber("000\\-00\\-0000", 123456789);
-		expect(result).toContain("-");
-	});
-
-	it("###,### for zero", () => {
-		expect(formatNumber("###,###", 0)).toBe("");
-	});
-});
-
-// ============================================================
-// Additional edge cases for deeper coverage
-// ============================================================
-describe("SSF: write_num_flt float paths", () => {
-	it("#,##0 for float", () => {
-		expect(formatNumber("#,##0", 1234.5)).toBe("1,235");
-	});
-
-	it("0.000 three fixed decimals", () => {
-		const result = formatNumber("0.000", 3.14);
-		expect(result).toBe("3.140");
-	});
-
-	it("#,##0.0 one decimal", () => {
-		expect(formatNumber("#,##0.0", 1234.56)).toBe("1,234.6");
-	});
-
-	it("# ?/? for float", () => {
-		const result = formatNumber("# ?/?", 2.75);
-		expect(result).toContain("/");
-	});
-
-	it("?/? for float without whole part", () => {
-		const result = formatNumber("?/?", 0.5);
-		expect(result).toContain("1/2");
-	});
-
-	it("0.00 for negative float", () => {
-		const result = formatNumber("0.00", -3.14);
-		expect(result).toBe("-3.14");
-	});
-
-	it("$0.00 for float", () => {
-		expect(formatNumber("$0.00", 42.5)).toBe("$42.50");
-	});
-
-	it("(0.00) parenthesized negative float", () => {
-		const result = formatNumber("(0.00)", -42.5);
-		expect(result).toContain("(");
-	});
-
-	it("(0.00) parenthesized positive float", () => {
-		const result = formatNumber("(0.00)", 42.5);
-		expect(result).toBeTruthy();
-	});
-});
-
-// ============================================================
-// write_num_exp edge cases
-// ============================================================
-describe("SSF: scientific notation edge cases", () => {
-	it("0.00E+00 zero", () => {
-		expect(formatNumber("0.00E+00", 0)).toBe("0.00E+00");
-	});
-
-	it("0.00E+00 negative", () => {
-		const result = formatNumber("0.00E+00", -12345);
-		expect(result).toMatch(/-1\.23E\+04/);
-	});
-
-	it("0E-00 strips plus sign from positive exponent", () => {
-		const result = formatNumber("0.00E-00", 12345);
-		// E- format: positive exponents should not show "+"
-		expect(result).not.toContain("E+");
-	});
-});
-
-// ============================================================
-// Date formats with subsecond precision
-// ============================================================
-describe("SSF: subsecond precision", () => {
-	it("ss.0 one decimal", () => {
-		const result = formatNumber("ss.0", 44197.500115);
-		expect(result).toMatch(/\d{2}\.\d/);
-	});
-
-	it("ss.000 three decimals", () => {
-		const result = formatNumber("ss.000", 44197.500115);
-		expect(result).toBeTruthy();
-	});
-
-	it("[hh]:mm:ss.00 exercises subsecond elapsed time", () => {
-		const result = formatNumber("[hh]:mm:ss.00", 1.5);
-		expect(result).toContain("36:00:");
-	});
-});
-
-// ============================================================
-// Format string splitting (SSF_split_fmt)
-// ============================================================
-describe("SSF: format string section selection", () => {
-	it("single section applies to all", () => {
-		expect(formatNumber("0.00", 42)).toBe("42.00");
-		expect(formatNumber("0.00", -42)).toBe("-42.00");
-		expect(formatNumber("0.00", 0)).toBe("0.00");
-	});
-
-	it("two sections: positive;negative", () => {
-		expect(formatNumber("0.00;(0.00)", 42)).toBe("42.00");
-		expect(formatNumber("0.00;(0.00)", -42)).toBe("(42.00)");
-		expect(formatNumber("0.00;(0.00)", 0)).toBe("0.00");
-	});
-
-	it("three sections: positive;negative;zero exercises all paths", () => {
-		const pos = formatNumber('0.00;(0.00);"zero"', 42);
-		expect(pos).toBeTruthy();
-		const neg = formatNumber('0.00;(0.00);"zero"', -42);
-		expect(neg).toContain("(");
-		const zero = formatNumber('0.00;(0.00);"zero"', 0);
-		expect(zero).toBe("zero");
-	});
-});
