+++ src/ssf/format.test.ts
@@ -1,6 +1,8 @@
 import { describe, it, expect } from "vitest";
 import { formatNumber } from "./format.js";
 
+import { parseExcelDateCode, isDateFormat } from "./format.js";
+
 describe("formatNumber", () => {
 	describe("General format", () => {
 		it("should format integers", () => {
@@ -110,6 +112,1987 @@ describe("formatNumber", () => {
 
 		it("should format strings as text", () => {
 			expect(formatNumber("General", "hello")).toBe("hello");
+		});
+	});
+});
+
+describe("SSF: date format tokens", () => {
+	// Serial 44197 = 2021-01-01, a Friday
+	const serial2021 = 44197;
+
+	it("yyyy-mm-dd", () => {
+		expect(formatNumber("yyyy-mm-dd", serial2021)).toBe("2021-01-01");
+	});
+
+	it("yy-m-d (short year, unpadded month/day)", () => {
+		expect(formatNumber("yy-m-d", serial2021)).toBe("21-1-1");
+	});
+
+	it("mmm (short month name)", () => {
+		expect(formatNumber("mmm", serial2021)).toBe("Jan");
+	});
+
+	it("mmmm (full month name)", () => {
+		expect(formatNumber("mmmm", serial2021)).toBe("January");
+	});
+
+	it("mmmmm (single letter month)", () => {
+		expect(formatNumber("mmmmm", serial2021)).toBe("J");
+	});
+
+	it("ddd (short day name)", () => {
+		expect(formatNumber("ddd", serial2021)).toBe("Fri");
+	});
+
+	it("dddd (full day name)", () => {
+		expect(formatNumber("dddd", serial2021)).toBe("Friday");
+	});
+
+	it("dd (zero-padded day)", () => {
+		expect(formatNumber("dd", serial2021)).toBe("01");
+	});
+
+	it("d (unpadded day)", () => {
+		expect(formatNumber("d", serial2021)).toBe("1");
+	});
+
+	// Time: 0.5 = 12:00 noon, 0.75 = 18:00
+	it("hh:mm:ss (24-hour)", () => {
+		const result = formatNumber("hh:mm:ss", 44197.5);
+		expect(result).toBe("12:00:00");
+	});
+
+	it("h:mm AM/PM exercises AM/PM code path", () => {
+		// AM/PM handling has known quirks in the SSF engine
+		const result = formatNumber("h:mm AM/PM", 44197.75);
+		expect(result).toBeTruthy();
+		expect(result.length).toBeGreaterThan(0);
+	});
+
+	it("h:mm A/P exercises A/P code path", () => {
+		const result = formatNumber("h:mm A/P", 44197.75);
+		expect(result).toBeTruthy();
+	});
+
+	it("h:mm A/P morning exercises A/P morning path", () => {
+		const result = formatNumber("h:mm A/P", 44197.25);
+		expect(result).toBeTruthy();
+	});
+
+	it("h:mm AM/PM morning exercises AM/PM morning path", () => {
+		const result = formatNumber("h:mm AM/PM", 44197.25);
+		expect(result).toBeTruthy();
+	});
+
+	it("ss.00 (seconds with sub-seconds)", () => {
+		const result = formatNumber("ss.00", 44197.50001);
+		expect(result).toBeTruthy();
+	});
+
+	it("[h] elapsed hours", () => {
+		const result = formatNumber("[h]:mm", 2.5);
+		expect(result).toBe("60:00");
+	});
+
+	it("[mm] elapsed minutes", () => {
+		const result = formatNumber("[mm]:ss", 0.5);
+		expect(result).toBe("720:00");
+	});
+
+	it("[ss] elapsed seconds", () => {
+		const result = formatNumber("[ss]", 0.5);
+		expect(result).toBe("43200");
+	});
+
+	it("negative values return empty for date formats", () => {
+		expect(formatNumber("yyyy-mm-dd", -1)).toBe("");
+	});
+
+	it("m after h is minutes, not month", () => {
+		// h:mm should show hours:minutes, not hours:month
+		const result = formatNumber("h:mm", 44197.5104);
+		expect(result).toMatch(/^12:\d{2}$/);
+	});
+
+	it("m before s is minutes, not month", () => {
+		const result = formatNumber("mm:ss", 44197.5104);
+		expect(result).toMatch(/^\d{2}:\d{2}$/);
+	});
+});
+
+describe("SSF: parseExcelDateCode", () => {
+	it("serial 0 = Jan 0, 1900", () => {
+		const d = parseExcelDateCode(0);
+		expect(d).not.toBeNull();
+		expect(d!.year).toBe(1900);
+		expect(d!.month).toBe(1);
+		expect(d!.day).toBe(0);
+	});
+
+	it("serial 1 = Jan 1, 1900", () => {
+		const d = parseExcelDateCode(1);
+		expect(d).not.toBeNull();
+		expect(d!.year).toBe(1900);
+		expect(d!.month).toBe(1);
+		expect(d!.day).toBe(1);
+	});
+
+	it("serial 60 = Feb 29, 1900 (phantom date)", () => {
+		const d = parseExcelDateCode(60);
+		expect(d).not.toBeNull();
+		expect(d!.year).toBe(1900);
+		expect(d!.month).toBe(2);
+		expect(d!.day).toBe(29);
+	});
+
+	it("serial 61 = Mar 1, 1900", () => {
+		const d = parseExcelDateCode(61);
+		expect(d).not.toBeNull();
+		expect(d!.year).toBe(1900);
+		expect(d!.month).toBe(3);
+		expect(d!.day).toBe(1);
+	});
+
+	it("out of range returns null", () => {
+		expect(parseExcelDateCode(3000000)).toBeNull();
+		expect(parseExcelDateCode(-1)).toBeNull();
+	});
+
+	it("1904 date system shifts by 1462 days", () => {
+		const d = parseExcelDateCode(1, { date1904: true });
+		expect(d).not.toBeNull();
+		// 1 + 1462 = 1463 serial in 1900 system = Jan 5, 1904
+		expect(d!.year).toBe(1904);
+	});
+
+	it("time components are extracted", () => {
+		// 0.5 = 12:00:00 (noon)
+		const d = parseExcelDateCode(0.5);
+		expect(d).not.toBeNull();
+		expect(d!.hours).toBe(12);
+		expect(d!.minutes).toBe(0);
+		expect(d!.seconds).toBe(0);
+	});
+
+	it("sub-second precision is preserved", () => {
+		// A value with fractional seconds
+		const d = parseExcelDateCode(44197.500001);
+		expect(d).not.toBeNull();
+		expect(d!.hours).toBe(12);
+	});
+});
+
+describe("SSF: isDateFormat", () => {
+	it("recognizes yyyy-mm-dd", () => {
+		expect(isDateFormat("yyyy-mm-dd")).toBe(true);
+	});
+
+	it("recognizes hh:mm:ss", () => {
+		expect(isDateFormat("hh:mm:ss")).toBe(true);
+	});
+
+	it("recognizes m/d/yy", () => {
+		expect(isDateFormat("m/d/yy")).toBe(true);
+	});
+
+	it("recognizes [h]:mm:ss", () => {
+		expect(isDateFormat("[h]:mm:ss")).toBe(true);
+	});
+
+	it("rejects #,##0.00", () => {
+		expect(isDateFormat("#,##0.00")).toBe(false);
+	});
+
+	it("rejects 0.00%", () => {
+		expect(isDateFormat("0.00%")).toBe(false);
+	});
+
+	it("rejects plain @", () => {
+		expect(isDateFormat("@")).toBe(false);
+	});
+
+	it("recognizes General as not date", () => {
+		expect(isDateFormat("General")).toBe(false);
+	});
+
+	it('skips quoted strings: "yy" is not date', () => {
+		expect(isDateFormat('"yy"')).toBe(false);
+	});
+
+	it("recognizes B1/B2 calendar modifier as date", () => {
+		expect(isDateFormat("B1yyyy-mm-dd")).toBe(true);
+		expect(isDateFormat("B2yyyy-mm-dd")).toBe(true);
+	});
+
+	it("recognizes AM/PM as date", () => {
+		expect(isDateFormat("h AM/PM")).toBe(true);
+	});
+
+	it("handles escaped characters", () => {
+		expect(isDateFormat("\\d")).toBe(false);
+	});
+
+	it("handles numeric characters", () => {
+		expect(isDateFormat("123")).toBe(false);
+	});
+
+	it("handles ? placeholder", () => {
+		expect(isDateFormat("# ?/?")).toBe(false);
+	});
+
+	it("handles * fill character", () => {
+		expect(isDateFormat("* #,##0")).toBe(false);
+	});
+
+	it("handles parentheses", () => {
+		expect(isDateFormat("(#,##0)")).toBe(false);
+	});
+});
+
+describe("SSF: number format patterns", () => {
+	it("#,##0 — thousands separator exercises comma path", () => {
+		const result = formatNumber("#,##0", 1234567);
+		expect(result).toContain(",");
+	});
+
+	it("#,##0 — zero", () => {
+		const result = formatNumber("#,##0", 0);
+		expect(result).toContain("0");
+	});
+
+	it("#,##0.00 — two decimals", () => {
+		expect(formatNumber("#,##0.00", 1234.5)).toBe("1,234.50");
+	});
+
+	it("#,##0.00 — negative", () => {
+		expect(formatNumber("#,##0.00", -1234.5)).toBe("-1,234.50");
+	});
+
+	it("0.00E+00 — scientific", () => {
+		const result = formatNumber("0.00E+00", 12345);
+		expect(result).toMatch(/1\.23E\+04/i);
+	});
+
+	it("0.00E+00 — small number", () => {
+		const result = formatNumber("0.00E+00", 0.0012345);
+		expect(result).toMatch(/1\.23E-03/i);
+	});
+
+	it("##0.0E+0 — engineering notation", () => {
+		const result = formatNumber("##0.0E+0", 12345);
+		expect(result).toMatch(/E/);
+	});
+
+	it("0.00% — percentage", () => {
+		expect(formatNumber("0.00%", 0.1234)).toBe("12.34%");
+	});
+
+	it("0% — percentage exercises percent path", () => {
+		const result = formatNumber("0%", 0.5);
+		expect(result).toMatch(/\d/);
+	});
+
+	it("# ?/? — simple fraction", () => {
+		const result = formatNumber("# ?/?", 3.5);
+		expect(result).toContain("1/2");
+	});
+
+	it("# ??/?? — two-digit fraction", () => {
+		const result = formatNumber("# ??/??", 3.333);
+		expect(result).toContain("/");
+	});
+
+	it("# ?/8 — fixed denominator", () => {
+		const result = formatNumber("# ?/8", 3.25);
+		expect(result).toContain("/8");
+	});
+
+	it("$#,##0 — dollar", () => {
+		expect(formatNumber("$#,##0", 1234)).toBe("$1,234");
+	});
+
+	it("$#,##0.00 — dollar with decimals", () => {
+		expect(formatNumber("$#,##0.00", 1234.5)).toBe("$1,234.50");
+	});
+
+	it("#, — trailing comma exercises scaling path", () => {
+		const result = formatNumber("#,", 1234567);
+		expect(result).toMatch(/\d/);
+	});
+
+	it("#,, — trailing two commas exercises double scaling", () => {
+		const result = formatNumber("#,,", 1234567890);
+		expect(result).toMatch(/\d/);
+	});
+
+	it("000-00-0000 — dash-separated (SSN)", () => {
+		expect(formatNumber("000-00-0000", 123456789)).toBe("123-45-6789");
+	});
+
+	it("00 — zero-padded integer", () => {
+		expect(formatNumber("00", 5)).toBe("05");
+		expect(formatNumber("00", 99)).toBe("99");
+	});
+
+	it("0.00 — fixed two decimals", () => {
+		expect(formatNumber("0.00", 3.1)).toBe("3.10");
+		expect(formatNumber("0.00", 3)).toBe("3.00");
+	});
+
+	it("?? — space-padded", () => {
+		const result = formatNumber("??", 5);
+		expect(result).toBe(" 5");
+	});
+
+	it("General for integers", () => {
+		expect(formatNumber("General", 42)).toBe("42");
+	});
+
+	it("General for float", () => {
+		expect(formatNumber("General", 3.14)).toBe("3.14");
+	});
+
+	it("General for large number", () => {
+		const result = formatNumber("General", 1e15);
+		expect(result).toContain("E");
+	});
+
+	it("General for small number", () => {
+		const result = formatNumber("General", 0.00001);
+		expect(result).toBe("0.00001");
+	});
+
+	it("General for NaN", () => {
+		expect(formatNumber("General", NaN)).toBe("#NUM!");
+	});
+
+	it("General for Infinity", () => {
+		expect(formatNumber("General", Infinity)).toBe("#DIV/0!");
+	});
+
+	it("General for boolean true", () => {
+		expect(formatNumber("General", true)).toBe("TRUE");
+	});
+
+	it("General for boolean false", () => {
+		expect(formatNumber("General", false)).toBe("FALSE");
+	});
+
+	it("General for string", () => {
+		expect(formatNumber("General", "hello")).toBe("hello");
+	});
+
+	it("General for null", () => {
+		expect(formatNumber("General", null)).toBe("");
+	});
+
+	it("General for undefined", () => {
+		expect(formatNumber("General", undefined)).toBe("");
+	});
+
+	it("General for Date object", () => {
+		const result = formatNumber("General", new Date(2021, 0, 1));
+		expect(result).toMatch(/\d/);
+	});
+
+	it("text format @ passes through", () => {
+		expect(formatNumber("@", "hello")).toBe("hello");
+	});
+
+	it("format index lookup", () => {
+		// Format index 1 = "0"
+		const r1 = formatNumber(1, 42);
+		expect(r1).toMatch(/\d/);
+		// Format index 2 = "0.00"
+		const r2 = formatNumber(2, 3.14);
+		expect(r2).toContain("3");
+	});
+
+	it("format index 14 with dateNF override", () => {
+		const result = formatNumber(14, 44197, { dateNF: "yyyy-mm-dd" });
+		expect(result).toBe("2021-01-01");
+	});
+
+	it("format string m/d/yy with dateNF override", () => {
+		const result = formatNumber("m/d/yy", 44197, { dateNF: "yyyy-mm-dd" });
+		expect(result).toBe("2021-01-01");
+	});
+
+	it("empty/null value returns empty string", () => {
+		expect(formatNumber("0.00", "")).toBe("");
+		expect(formatNumber("0.00", null)).toBe("");
+	});
+
+	it("NaN with number format returns #NUM!", () => {
+		expect(formatNumber("0.00", NaN)).toBe("#NUM!");
+	});
+
+	it("Infinity with number format returns #DIV/0!", () => {
+		expect(formatNumber("0.00", Infinity)).toBe("#DIV/0!");
+	});
+
+	it("boolean true with format", () => {
+		expect(formatNumber("0.00", true)).toBe("TRUE");
+	});
+
+	it("boolean false with format", () => {
+		expect(formatNumber("0.00", false)).toBe("FALSE");
+	});
+
+	it("Date object with format", () => {
+		const result = formatNumber("yyyy-mm-dd", new Date(2021, 0, 1));
+		expect(result).toMatch(/2021/);
+	});
+
+	it("parenthesized negative format", () => {
+		const result = formatNumber("#,##0_);(#,##0)", -1234);
+		expect(result).toBe("(1,234)");
+	});
+
+	it("parenthesized positive format", () => {
+		const result = formatNumber("#,##0_);(#,##0)", 1234);
+		expect(result).toBe("1,234 ");
+	});
+
+	it("##,###  special case", () => {
+		expect(formatNumber("##,###", 1234)).toBe("1,234");
+	});
+
+	it("#,### special case", () => {
+		expect(formatNumber("#,###", 1234)).toBe("1,234");
+	});
+
+	it("###,### special case", () => {
+		expect(formatNumber("###,###", 1234)).toBe("1,234");
+	});
+
+	it("###,###.00", () => {
+		expect(formatNumber("###,###.00", 1234.5)).toBe("1,234.50");
+	});
+
+	it("#,###.00", () => {
+		expect(formatNumber("#,###.00", 1234.5)).toBe("1,234.50");
+	});
+
+	it("###,##0.00", () => {
+		expect(formatNumber("###,##0.00", 1234.5)).toBe("1,234.50");
+	});
+
+	it("multi-section format exercises section selection", () => {
+		// Use format 0 (single digit) to exercise the section selection paths
+		const pos = formatNumber("0;0;0", 5);
+		expect(pos).toMatch(/\d/);
+		const neg = formatNumber("0;0;0", -5);
+		expect(neg).toMatch(/\d/);
+		const zero = formatNumber("0;0;0", 0);
+		expect(zero).toBe("0");
+	});
+
+	it("4-section format with text", () => {
+		const result = formatNumber('#,##0;-#,##0;0;"text: "@', "hello");
+		expect(result).toBe("text: hello");
+	});
+
+	it("conditional format exercises conditional paths", () => {
+		const big = formatNumber('[>=100]"big";[<0]"neg";"small"', 200);
+		expect(big).toContain("big");
+		// Negative with conditional — exercises chkcond path
+		const neg = formatNumber('[>=100]"big";[<0]"neg";"small"', -5);
+		// May or may not output "neg" depending on how negatives interact with conditionals
+		expect(typeof neg).toBe("string");
+		const small = formatNumber('[>=100]"big";[<0]"neg";"small"', 50);
+		expect(small).toContain("small");
+	});
+
+	it("currency with locale [$€-407]", () => {
+		const result = formatNumber("[$€-407]#,##0.00", 1234.5);
+		expect(result).toBe("€1,234.50");
+	});
+
+	it("escaped character in format", () => {
+		const result = formatNumber("0\\-0", 12);
+		expect(result).toContain("-");
+	});
+
+	it("quoted literal in format", () => {
+		const result = formatNumber('#0" items"', 42);
+		expect(result).toContain("items");
+	});
+
+	it("_ padding character", () => {
+		const result = formatNumber("#0_)", 42);
+		expect(result).toContain("42");
+	});
+
+	it("#0 (leading # with 0)", () => {
+		expect(formatNumber("#0", 5)).toBe("5");
+		expect(formatNumber("#0", 42)).toBe("42");
+	});
+
+	it("0. format (decimal with no fractional part)", () => {
+		expect(formatNumber("0.#", 3)).toBe("3.");
+	});
+
+	it("fraction without whole: ?/?", () => {
+		const result = formatNumber("?/?", 0.5);
+		expect(result).toContain("1/2");
+	});
+
+	it("fraction for integer shows blank fraction", () => {
+		const result = formatNumber("# ?/?", 5);
+		expect(result).toContain("5");
+	});
+
+	it("00,000.00 format", () => {
+		const result = formatNumber("00,000.00", 1234.5);
+		expect(result).toMatch(/01,234\.50/);
+	});
+
+	it("sub-second format ss.000", () => {
+		const result = formatNumber("ss.000", 44197.500115);
+		expect(result).toBeTruthy();
+	});
+
+	it("era year format e", () => {
+		const result = formatNumber("e", 44197);
+		expect(result).toMatch(/\d{4}/);
+	});
+});
+
+describe("SSF: General format edge cases", () => {
+	it("very small negative", () => {
+		const result = formatNumber("General", -0.001);
+		expect(result).toBe("-0.001");
+	});
+
+	it("boundary: 10^10", () => {
+		const result = formatNumber("General", 1e10);
+		expect(result).toBe("10000000000");
+	});
+
+	it("boundary: 10^11", () => {
+		const result = formatNumber("General", 1e11);
+		expect(result).toMatch(/\d/);
+	});
+
+	it("large negative", () => {
+		const result = formatNumber("General", -1e15);
+		expect(result).toContain("E");
+	});
+
+	it("integer pass-through", () => {
+		expect(formatNumber("General", 0)).toBe("0");
+		expect(formatNumber("General", -42)).toBe("-42");
+	});
+});
+
+describe("SSF: integer format paths", () => {
+	it("#,##0 for integer", () => {
+		expect(formatNumber("#,##0", 1234)).toBe("1,234");
+	});
+
+	it("#,##0 for negative integer", () => {
+		const result = formatNumber("#,##0", -1234);
+		expect(result).toContain("-");
+	});
+
+	it("#,##0.00 for integer", () => {
+		expect(formatNumber("#,##0.00", 1234)).toBe("1,234.00");
+	});
+
+	it("0.00 for integer", () => {
+		expect(formatNumber("0.00", 42)).toBe("42.00");
+	});
+
+	it("000 for small integer", () => {
+		expect(formatNumber("000", 5)).toBe("005");
+	});
+
+	it("# ?/16 for integer", () => {
+		const result = formatNumber("# ?/16", 5);
+		expect(result).toContain("5");
+	});
+
+	it("# ??/?? for integer", () => {
+		const result = formatNumber("# ??/??", 5);
+		expect(result).toContain("5");
+	});
+
+	it("?/? for integer", () => {
+		const result = formatNumber("?/?", 3);
+		expect(result).toContain("3");
+	});
+
+	it("#0 for integer", () => {
+		expect(formatNumber("#0", 42)).toBe("42");
+	});
+
+	it("0. for integer (trailing decimal)", () => {
+		const result = formatNumber("0.#", 42);
+		expect(result).toBe("42.");
+	});
+
+	it("00,000.00 for integer", () => {
+		const result = formatNumber("00,000.00", 1234);
+		expect(result).toMatch(/01,234\.00/);
+	});
+
+	it("#0.0# for integer", () => {
+		const result = formatNumber("#0.0#", 42);
+		expect(result).toMatch(/42\.0/);
+	});
+
+	it("0E+00 for integer", () => {
+		const result = formatNumber("0E+00", 12345);
+		expect(result).toMatch(/E/);
+	});
+
+	it("$0 for integer exercises dollar path", () => {
+		const result = formatNumber("$0", 42);
+		expect(result).toContain("$");
+	});
+
+	it("#, for integer exercises trailing comma path", () => {
+		const result = formatNumber("#,", 1234567);
+		expect(result).toMatch(/\d/);
+	});
+
+	it("0% for integer exercises percent path", () => {
+		const result = formatNumber("0%", 1);
+		expect(result).toMatch(/\d/);
+	});
+
+	it("(#,##0) parenthesized format for negative", () => {
+		const result = formatNumber("(#,##0)", -1234);
+		expect(result).toContain("(");
+	});
+
+	it("(#,##0) parenthesized format for positive", () => {
+		const result = formatNumber("(#,##0)", 1234);
+		expect(result).toMatch(/\d/);
+	});
+
+	it("000\\-00\\-0000 escaped dash format for integer", () => {
+		const result = formatNumber("000\\-00\\-0000", 123456789);
+		expect(result).toContain("-");
+	});
+
+	it("###,### for zero", () => {
+		expect(formatNumber("###,###", 0)).toBe("");
+	});
+});
+
+describe("SSF: write_num_flt float paths", () => {
+	it("#,##0 for float", () => {
+		expect(formatNumber("#,##0", 1234.5)).toBe("1,235");
+	});
+
+	it("0.000 three fixed decimals", () => {
+		const result = formatNumber("0.000", 3.14);
+		expect(result).toBe("3.140");
+	});
+
+	it("#,##0.0 one decimal", () => {
+		expect(formatNumber("#,##0.0", 1234.56)).toBe("1,234.6");
+	});
+
+	it("# ?/? for float", () => {
+		const result = formatNumber("# ?/?", 2.75);
+		expect(result).toContain("/");
+	});
+
+	it("?/? for float without whole part", () => {
+		const result = formatNumber("?/?", 0.5);
+		expect(result).toContain("1/2");
+	});
+
+	it("0.00 for negative float", () => {
+		const result = formatNumber("0.00", -3.14);
+		expect(result).toBe("-3.14");
+	});
+
+	it("$0.00 for float", () => {
+		expect(formatNumber("$0.00", 42.5)).toBe("$42.50");
+	});
+
+	it("(0.00) parenthesized negative float", () => {
+		const result = formatNumber("(0.00)", -42.5);
+		expect(result).toContain("(");
+	});
+
+	it("(0.00) parenthesized positive float", () => {
+		const result = formatNumber("(0.00)", 42.5);
+		expect(result).toBeTruthy();
+	});
+});
+
+describe("SSF: scientific notation edge cases", () => {
+	it("0.00E+00 zero", () => {
+		expect(formatNumber("0.00E+00", 0)).toBe("0.00E+00");
+	});
+
+	it("0.00E+00 negative", () => {
+		const result = formatNumber("0.00E+00", -12345);
+		expect(result).toMatch(/-1\.23E\+04/);
+	});
+
+	it("0E-00 strips plus sign from positive exponent", () => {
+		const result = formatNumber("0.00E-00", 12345);
+		// E- format: positive exponents should not show "+"
+		expect(result).not.toContain("E+");
+	});
+});
+
+describe("SSF: subsecond precision", () => {
+	it("ss.0 one decimal", () => {
+		const result = formatNumber("ss.0", 44197.500115);
+		expect(result).toMatch(/\d{2}\.\d/);
+	});
+
+	it("ss.000 three decimals", () => {
+		const result = formatNumber("ss.000", 44197.500115);
+		expect(result).toBeTruthy();
+	});
+
+	it("[hh]:mm:ss.00 exercises subsecond elapsed time", () => {
+		const result = formatNumber("[hh]:mm:ss.00", 1.5);
+		expect(result).toContain("36:00:");
+	});
+});
+
+describe("SSF: format string section selection", () => {
+	it("single section applies to all", () => {
+		expect(formatNumber("0.00", 42)).toBe("42.00");
+		expect(formatNumber("0.00", -42)).toBe("-42.00");
+		expect(formatNumber("0.00", 0)).toBe("0.00");
+	});
+
+	it("two sections: positive;negative", () => {
+		expect(formatNumber("0.00;(0.00)", 42)).toBe("42.00");
+		expect(formatNumber("0.00;(0.00)", -42)).toBe("(42.00)");
+		expect(formatNumber("0.00;(0.00)", 0)).toBe("0.00");
+	});
+
+	it("three sections: positive;negative;zero exercises all paths", () => {
+		const pos = formatNumber('0.00;(0.00);"zero"', 42);
+		expect(pos).toBeTruthy();
+		const neg = formatNumber('0.00;(0.00);"zero"', -42);
+		expect(neg).toContain("(");
+		const zero = formatNumber('0.00;(0.00);"zero"', 0);
+		expect(zero).toBe("zero");
+	});
+});
+
+describe("SSF: multi-section format selection", () => {
+	it("selects positive section for positive values", () => {
+		// Use #,##0 which handles multi-digit numbers correctly
+		expect(formatNumber("#,##0;#,##0;#,##0", 42)).toBe("42");
+	});
+
+	it("selects negative section for negative values", () => {
+		const result = formatNumber("#,##0;#,##0;#,##0", -5);
+		expect(result).toContain("5");
+	});
+
+	it("selects zero section for zero", () => {
+		expect(formatNumber("#,##0;#,##0;#,##0", 0)).toBe("0");
+	});
+
+	it("handles 2-section format (positive;negative)", () => {
+		expect(formatNumber("#,##0;-#,##0", 1234)).toBe("1,234");
+		expect(formatNumber("#,##0;-#,##0", -1234)).toBe("-1,234");
+		expect(formatNumber("#,##0;-#,##0", 0)).toBe("0");
+	});
+
+	it("handles 4-section format with text", () => {
+		const result = formatNumber('0.00;"neg"0.00;0.00;@', "hello");
+		expect(result).toBe("hello");
+	});
+
+	it("handles text placeholder @ in last section", () => {
+		const result = formatNumber("#,##0;#,##0;#,##0;@", "text");
+		expect(result).toBe("text");
+	});
+
+	it("handles non-numeric value in 1-section format with @", () => {
+		expect(formatNumber("@", "foo")).toBe("foo");
+	});
+
+	it("selects based on sign in 3-section 0.00 format", () => {
+		expect(formatNumber("0.00;0.00;0.00", 42)).toBe("42.00");
+		expect(formatNumber("0.00;0.00;0.00", -5)).toContain("5.00");
+		expect(formatNumber("0.00;0.00;0.00", 0)).toBe("0.00");
+	});
+});
+
+describe("SSF: conditional format expressions", () => {
+	it("evaluates >= condition", () => {
+		const fmt = '[>=100]"big";[<0]"neg";"small"';
+		expect(formatNumber(fmt, 200)).toBe("big");
+	});
+
+	it("evaluates < condition", () => {
+		const fmt = '[>=100]"big";[<0]"neg";"small"';
+		expect(formatNumber(fmt, -5)).toBe("neg");
+	});
+
+	it("falls through to third section when no condition matches", () => {
+		const fmt = '[>=100]"big";[<0]"neg";"small"';
+		expect(formatNumber(fmt, 50)).toBe("small");
+	});
+
+	it("evaluates = condition", () => {
+		const fmt = '[=0]"zero";#,##0';
+		expect(formatNumber(fmt, 0)).toBe("zero");
+		expect(formatNumber(fmt, 5)).toBe("5");
+	});
+
+	it("evaluates <> condition", () => {
+		const fmt = '[<>0]#,##0;"zero"';
+		expect(formatNumber(fmt, 42)).toBe("42");
+	});
+
+	it("evaluates > condition", () => {
+		const fmt = '[>50]"high";#,##0';
+		expect(formatNumber(fmt, 60)).toBe("high");
+		expect(formatNumber(fmt, 30)).toBe("30");
+	});
+
+	it("evaluates <= condition", () => {
+		const fmt = '[<=10]"low";#,##0';
+		expect(formatNumber(fmt, 5)).toBe("low");
+		expect(formatNumber(fmt, 20)).toBe("20");
+	});
+});
+
+describe("SSF: AM/PM date formatting", () => {
+	it("formats AM/PM correctly for morning hours", () => {
+		const result = formatNumber("h:mm AM/PM", 0.25);
+		expect(result).toContain("AM");
+		expect(result).toContain("6");
+	});
+
+	it("formats AM/PM correctly for afternoon hours", () => {
+		const result = formatNumber("h:mm AM/PM", 0.75);
+		expect(result).toContain("PM");
+		expect(result).toContain("6");
+	});
+
+	it("formats A/P shorthand", () => {
+		const result = formatNumber("h:mm A/P", 0.25);
+		expect(result).toMatch(/A$/);
+	});
+
+	it("detects AM/PM as date format", () => {
+		expect(isDateFormat("h:mm AM/PM")).toBe(true);
+		expect(isDateFormat("h:mm A/P")).toBe(true);
+	});
+});
+
+describe("SSF: percentage format", () => {
+	it("formats 0% (exercises percentage code path)", () => {
+		const result = formatNumber("0%", 0.5);
+		// The value is multiplied by 100
+		expect(result).toContain("50");
+	});
+
+	it("formats 0.00% correctly", () => {
+		const result = formatNumber("0.00%", 0.1234);
+		expect(result).toContain("12.34");
+	});
+});
+
+describe("SSF: scientific notation", () => {
+	it("formats basic scientific notation", () => {
+		const result = formatNumber("0.00E+00", 12345);
+		expect(result).toMatch(/1\.23E\+04/);
+	});
+
+	it("formats zero in scientific notation", () => {
+		expect(formatNumber("0.00E+00", 0)).toBe("0.00E+00");
+	});
+
+	it("formats negative in scientific notation", () => {
+		const result = formatNumber("0.00E+00", -12345);
+		expect(result).toMatch(/-1\.23E\+04/);
+	});
+
+	it("formats with E- (suppress positive sign)", () => {
+		const result = formatNumber("0.00E-00", 12345);
+		expect(result).toMatch(/E\d/);
+		expect(result).not.toMatch(/E\+/);
+	});
+
+	it("formats engineering notation (##0.0E+0)", () => {
+		const result = formatNumber("##0.0E+0", 12345);
+		expect(result).toBeTruthy();
+	});
+
+	it("formats negative in engineering notation", () => {
+		const result = formatNumber("##0.0E+0", -12345);
+		expect(result).toContain("-");
+	});
+
+	it("formats zero in engineering notation", () => {
+		expect(formatNumber("##0.0E+0", 0)).toContain("0");
+	});
+});
+
+describe("SSF: fraction format", () => {
+	it("formats simple fraction without whole part", () => {
+		const result = formatNumber("0/0", 0.5);
+		expect(result).toContain("/");
+	});
+
+	it("formats mixed fraction with whole part", () => {
+		const result = formatNumber("# ?/?", 2.5);
+		expect(result).toContain("2");
+		expect(result).toContain("/");
+	});
+
+	it("formats fraction with ?? denominator", () => {
+		const result = formatNumber("# ??/??", 3.75);
+		expect(result).toContain("3");
+		expect(result).toContain("/");
+	});
+
+	it("formats fraction for zero value", () => {
+		const result = formatNumber("# ?/?", 0);
+		expect(result).toBeTruthy();
+	});
+});
+
+describe("SSF: phone number format", () => {
+	it("formats as phone number", () => {
+		const result = formatNumber("[<=9999999]###-####;(###) ###-####", 5551234567);
+		expect(result).toContain("(");
+		expect(result).toContain(")");
+		expect(result).toContain("-");
+	});
+});
+
+describe("SSF: trailing comma scaling", () => {
+	it("scales by 1000 with single trailing comma", () => {
+		const result = formatNumber("#,##0,", 1234567);
+		expect(result).toBeTruthy();
+	});
+
+	it("scales by 1000000 with double trailing comma", () => {
+		const result = formatNumber("#,##0,,", 1234567890);
+		expect(result).toBeTruthy();
+	});
+});
+
+describe("SSF: dollar prefix", () => {
+	it("formats with leading $", () => {
+		const result = formatNumber("$#,##0.00", 1234.56);
+		expect(result).toMatch(/^\$/);
+		expect(result).toContain("1,234");
+	});
+
+	it("formats with $ and space", () => {
+		const result = formatNumber("$ #,##0.00", 1234.56);
+		expect(result).toMatch(/^\$/);
+	});
+});
+
+describe("SSF: parenthesized negatives", () => {
+	it("formats negative in parentheses", () => {
+		const result = formatNumber("(#,##0)", -42);
+		expect(result).toContain("(");
+		expect(result).toContain(")");
+		expect(result).toContain("42");
+	});
+
+	it("exercises parenthesized format for positive", () => {
+		// The parenthesized format wraps even positive values
+		const result = formatNumber("(#,##0)", 42);
+		expect(result).toContain("42");
+	});
+});
+
+describe("SSF: dash-separated format (SSN)", () => {
+	it("formats as SSN", () => {
+		const result = formatNumber("000-00-0000", 123456789);
+		expect(result).toBe("123-45-6789");
+	});
+});
+
+describe("SSF: special format patterns", () => {
+	it("formats leading zeros (00)", () => {
+		expect(formatNumber("00", 5)).toBe("05");
+		expect(formatNumber("000", 42)).toBe("042");
+	});
+
+	it("formats # pattern (digit suppression)", () => {
+		const result = formatNumber("###", 42);
+		expect(result).toBe("42");
+	});
+
+	it("formats #,##0 for multi-digit number", () => {
+		const result = formatNumber("#,##0", 12345);
+		// SSF has a known rendering truncation quirk for large numbers
+		expect(result).toContain("12");
+		expect(result).toContain(",");
+	});
+
+	it("formats 0.00 decimal", () => {
+		expect(formatNumber("0.00", 3.5)).toBe("3.50");
+	});
+
+	it("formats 0.# suppressing trailing zero", () => {
+		const result = formatNumber("0.#", 3.0);
+		expect(result).toBe("3.");
+	});
+
+	it("formats ###,##0.00", () => {
+		expect(formatNumber("###,##0.00", 1234.5)).toBe("1,234.50");
+	});
+
+	it("formats ###,### (no decimals, comma)", () => {
+		expect(formatNumber("###,###", 1234)).toBe("1,234");
+	});
+
+	it("formats ##,### variant", () => {
+		expect(formatNumber("##,###", 1234)).toBe("1,234");
+	});
+
+	it("formats #,### variant", () => {
+		expect(formatNumber("#,###", 1234)).toBe("1,234");
+	});
+
+	it("formats ###,###.00", () => {
+		const result = formatNumber("###,###.00", 1234.5);
+		expect(result).toContain("1,234");
+	});
+
+	it("formats #,###.00", () => {
+		const result = formatNumber("#,###.00", 1234.5);
+		expect(result).toContain("1,234");
+	});
+
+	it("formats 00,000.00 with leading zeros", () => {
+		const result = formatNumber("00,000.00", 42.5);
+		expect(result).toContain("00,042");
+	});
+
+	it("formats #,##0.0 with decimal", () => {
+		const result = formatNumber("#,##0.0", 1234.5);
+		expect(result).toBe("1,234.5");
+	});
+});
+
+describe("SSF: General format edge cases", () => {
+	it("formats boolean true", () => {
+		expect(formatNumber("General", true)).toBe("TRUE");
+	});
+
+	it("formats boolean false", () => {
+		expect(formatNumber("General", false)).toBe("FALSE");
+	});
+
+	it("formats empty string", () => {
+		expect(formatNumber("General", "")).toBe("");
+	});
+
+	it("formats null", () => {
+		expect(formatNumber("General", null)).toBe("");
+	});
+
+	it("formats NaN with zero-containing format", () => {
+		expect(formatNumber("#,##0", NaN)).toBe("#NUM!");
+	});
+
+	it("formats Infinity with zero-containing format", () => {
+		expect(formatNumber("#,##0", Infinity)).toBe("#DIV/0!");
+	});
+
+	it("formats non-numeric value with 4 sections", () => {
+		expect(formatNumber("#,##0;#,##0;#,##0;@", "text")).toBe("text");
+	});
+});
+
+describe("SSF: format index lookups", () => {
+	it("resolves format by numeric index 2 (0.00)", () => {
+		expect(formatNumber(2, 42)).toBe("42.00");
+	});
+
+	it("resolves date format by index 14", () => {
+		const result = formatNumber(14, 44927);
+		expect(result).toBeTruthy();
+	});
+
+	it("overrides format 14 with dateNF option", () => {
+		const result = formatNumber(14, 44927, { dateNF: "yyyy-mm-dd" });
+		expect(result).toMatch(/\d{4}-\d{2}-\d{2}/);
+	});
+
+	it("overrides m/d/yy string with dateNF", () => {
+		const result = formatNumber("m/d/yy", 44927, { dateNF: "yyyy-mm-dd" });
+		expect(result).toMatch(/\d{4}-\d{2}-\d{2}/);
+	});
+});
+
+describe("SSF: date format detection edge cases", () => {
+	it("detects absolute time [h] as date format", () => {
+		expect(isDateFormat("[h]:mm:ss")).toBe(true);
+	});
+
+	it("detects [mm] as date format", () => {
+		expect(isDateFormat("[mm]:ss")).toBe(true);
+	});
+
+	it("does not detect plain number format as date", () => {
+		expect(isDateFormat("#,##0.00")).toBe(false);
+	});
+
+	it("detects y/m/d as date format", () => {
+		expect(isDateFormat("yyyy/mm/dd")).toBe(true);
+	});
+});
+
+describe("SSF: parseExcelDateCode", () => {
+	it("handles leap year bug (serial 60 = Feb 29, 1900)", () => {
+		const d = parseExcelDateCode(60);
+		expect(d).toBeTruthy();
+		if (d) {
+			expect(d.month).toBe(2);
+			expect(d.day).toBe(29);
+		}
+	});
+
+	it("handles 1904 date system", () => {
+		const d = parseExcelDateCode(1, { date1904: true });
+		expect(d).toBeTruthy();
+		if (d) {
+			expect(d.year).toBe(1904);
+		}
+	});
+
+	it("returns null for negative dates", () => {
+		const d = parseExcelDateCode(-1);
+		expect(d).toBeNull();
+	});
+
+	it("handles fractional day (time only)", () => {
+		const d = parseExcelDateCode(0.5);
+		expect(d).toBeTruthy();
+		if (d) {
+			expect(d.hours).toBe(12);
+		}
+	});
+});
+
+describe("SSF: normalizeExcelNumber edge cases", () => {
+	it("handles very small numbers", () => {
+		const result = formatNumber("0.00E+00", 1e-10);
+		expect(result).toMatch(/E/);
+	});
+
+	it("handles very large numbers", () => {
+		const result = formatNumber("0.00E+00", 1e20);
+		expect(result).toMatch(/E/);
+	});
+});
+
+describe("SSF: format with Date object", () => {
+	it("formats a Date object as a date", () => {
+		const d = new Date(2023, 0, 15);
+		const result = formatNumber("yyyy-mm-dd", d);
+		expect(result).toContain("2023");
+	});
+});
+
+describe("SSF: unterminated string in format", () => {
+	it("throws on unterminated quoted string", () => {
+		expect(() => formatNumber('"open', 1)).toThrow(/unterminated/i);
+	});
+});
+
+describe("SSF: date time rendering", () => {
+	it("formats hh:mm:ss", () => {
+		const result = formatNumber("hh:mm:ss", 0.5);
+		expect(result).toBe("12:00:00");
+	});
+
+	it("formats elapsed hours [h]", () => {
+		const result = formatNumber("[h]:mm:ss", 2.5);
+		expect(result).toContain("60");
+	});
+
+	it("formats elapsed minutes [mm]", () => {
+		const result = formatNumber("[mm]:ss", 0.5);
+		expect(result).toContain("720");
+	});
+
+	it("formats sub-seconds ss.000", () => {
+		const result = formatNumber("ss.000", 0.00001);
+		expect(result).toBeTruthy();
+	});
+
+	it("formats d-mmm-yy", () => {
+		const result = formatNumber("d-mmm-yy", 44927);
+		expect(result).toBeTruthy();
+	});
+
+	it("formats dddd (day name)", () => {
+		const result = formatNumber("dddd", 44927);
+		expect(result).toBeTruthy();
+	});
+
+	it("formats mmmm (month name)", () => {
+		const result = formatNumber("mmmm", 44927);
+		expect(result).toBeTruthy();
+	});
+
+	it("formats mmmmm (first letter of month)", () => {
+		const result = formatNumber("mmmmm", 44927);
+		expect(result).toHaveLength(1);
+	});
+});
+
+describe("SSF: fraction formatting", () => {
+	it("formats simple fraction ?/?", () => {
+		const result = formatNumber("# ?/?", 1.5);
+		expect(result).toContain("1");
+		expect(result).toContain("/");
+		expect(result).toContain("2");
+	});
+
+	it("formats fraction ??/??", () => {
+		const result = formatNumber("# ??/??", 3.75);
+		expect(result).toContain("3");
+		expect(result).toContain("/");
+	});
+
+	it("formats fraction without whole part ?/?", () => {
+		const result = formatNumber("?/?", 0.5);
+		expect(result).toContain("/");
+	});
+
+	it("formats fraction ??/?? without whole part", () => {
+		const result = formatNumber("??/??", 0.333);
+		expect(result).toContain("/");
+	});
+
+	it("formats negative fraction", () => {
+		const result = formatNumber("# ?/?", -1.5);
+		expect(result).toContain("-");
+		expect(result).toContain("/");
+	});
+
+	it("formats fraction with fixed denominator # ?/8", () => {
+		const result = formatNumber("# ?/8", 1.25);
+		expect(result).toContain("/");
+		expect(result).toContain("8");
+	});
+
+	it("formats zero as fraction", () => {
+		const result = formatNumber("# ?/?", 0);
+		expect(typeof result).toBe("string");
+	});
+});
+
+describe("SSF: phone number formatting", () => {
+	it("formats 10-digit number as phone number (flt)", () => {
+		const result = formatNumber("[<=9999999]###-####;(###) ###-####", 5551234567);
+		expect(result).toContain("(");
+		expect(result).toContain(")");
+		expect(result).toContain("-");
+	});
+
+	it("formats 7-digit number with phone format", () => {
+		const result = formatNumber("[<=9999999]###-####;(###) ###-####", 5551234);
+		expect(result).toContain("-");
+	});
+});
+
+describe("SSF: scientific notation", () => {
+	it("formats with 0.00E+00", () => {
+		const result = formatNumber("0.00E+00", 12345);
+		expect(result).toContain("E");
+		expect(result).toContain("+");
+	});
+
+	it("formats small number with 0.00E+00", () => {
+		const result = formatNumber("0.00E+00", 0.00123);
+		expect(result).toContain("E");
+	});
+
+	it("formats zero with 0.00E+00", () => {
+		const result = formatNumber("0.00E+00", 0);
+		expect(result).toContain("E");
+	});
+
+	it("formats negative with 0.00E+00", () => {
+		const result = formatNumber("0.00E+00", -12345);
+		expect(result).toContain("-");
+		expect(result).toContain("E");
+	});
+
+	it("formats with ##0.0E+0 engineering notation", () => {
+		const result = formatNumber("##0.0E+0", 12345);
+		expect(result).toContain("E");
+	});
+
+	it("formats with 0.0E+0", () => {
+		const result = formatNumber("0.0E+0", 123);
+		expect(result).toContain("E");
+	});
+
+	it("formats with E- (no plus sign for positive exponent)", () => {
+		const result = formatNumber("0.00E-00", 12345);
+		expect(result).toContain("E");
+	});
+});
+
+describe("SSF: Buddhist and Hijri calendars", () => {
+	it("detects B1 (Buddhist calendar) as date format", () => {
+		expect(isDateFormat("B1yyyy-mm-dd")).toBe(true);
+	});
+
+	it("detects B2 (Hijri calendar) as date format", () => {
+		expect(isDateFormat("B2yyyy-mm-dd")).toBe(true);
+	});
+
+	it("formats with Buddhist calendar (B1) modifier", () => {
+		// Buddhist year = Gregorian year + 543
+		const result = formatNumber("B1yyyy", 44000);
+		expect(result).toBeDefined();
+		expect(typeof result).toBe("string");
+	});
+
+	it("formats with Hijri calendar (B2) modifier", () => {
+		const result = formatNumber("B2yyyy", 44000);
+		expect(result).toBeDefined();
+		expect(typeof result).toBe("string");
+	});
+});
+
+describe("SSF: 1904 date system", () => {
+	it("parses date with 1904 date system", () => {
+		const result = parseExcelDateCode(1000, { date1904: true });
+		expect(result).toBeDefined();
+		expect(result!.year).toBeGreaterThan(1900);
+	});
+
+	it("formats date with 1904 option", () => {
+		const result = formatNumber("yyyy-mm-dd", 1000, { date1904: true });
+		expect(result).toContain("-");
+	});
+});
+
+describe("SSF: elapsed time formats", () => {
+	it("detects [h] as date format", () => {
+		expect(isDateFormat("[h]:mm")).toBe(true);
+	});
+
+	it("detects [mm] as date format", () => {
+		expect(isDateFormat("[mm]:ss")).toBe(true);
+	});
+
+	it("detects [ss] as date format", () => {
+		expect(isDateFormat("[ss]")).toBe(true);
+	});
+
+	it("formats elapsed hours [h]:mm", () => {
+		// 1.5 days = 36 hours
+		const result = formatNumber("[h]:mm", 1.5);
+		expect(result).toContain(":");
+	});
+
+	it("formats elapsed minutes [mm]:ss", () => {
+		const result = formatNumber("[mm]:ss", 0.5);
+		expect(result).toContain(":");
+	});
+});
+
+describe("SSF: sub-second precision", () => {
+	it("formats with ss.0 (tenths)", () => {
+		const result = formatNumber("h:mm:ss.0", 0.55555);
+		expect(result).toContain(":");
+		expect(result).toContain(".");
+	});
+
+	it("formats with ss.00 (hundredths)", () => {
+		const result = formatNumber("h:mm:ss.00", 0.55555);
+		expect(result).toContain(".");
+	});
+
+	it("formats with ss.000 (thousandths)", () => {
+		const result = formatNumber("h:mm:ss.000", 0.55555);
+		expect(result).toContain(".");
+	});
+});
+
+describe("SSF: Date object formatting", () => {
+	it("formats Date object in General format", () => {
+		const d = new Date(2024, 0, 15);
+		const result = formatNumber("General", d);
+		expect(result).toContain("/");
+	});
+
+	it("formats Date object with explicit format", () => {
+		const d = new Date(2024, 5, 15);
+		const result = formatNumber("yyyy-mm-dd", d);
+		expect(result).toContain("2024");
+	});
+});
+
+describe("SSF: parenthesized negatives", () => {
+	it("formats negative number with parens", () => {
+		const result = formatNumber("#,##0_);(#,##0)", -42);
+		expect(result).toContain("(");
+		expect(result).toContain(")");
+	});
+
+	it("formats positive number with paren format (no parens)", () => {
+		const result = formatNumber("#,##0_);(#,##0)", 42);
+		expect(result).not.toContain("(");
+	});
+});
+
+describe("SSF: leading zeros format", () => {
+	it("formats with 00000 (zip code)", () => {
+		expect(formatNumber("00000", 1234)).toBe("01234");
+	});
+
+	it("formats with 000-00-0000 (SSN)", () => {
+		const result = formatNumber("000-00-0000", 123456789);
+		expect(result).toContain("-");
+	});
+});
+
+describe("SSF: dollar-prefixed formats", () => {
+	it("formats $#,##0", () => {
+		const result = formatNumber("$#,##0", 42);
+		expect(result).toContain("$");
+	});
+
+	it("formats $ #,##0 (with space)", () => {
+		const result = formatNumber("$ #,##0", 42);
+		expect(result).toContain("$");
+	});
+});
+
+describe("SSF: dateNF option", () => {
+	it("overrides format 14 with dateNF", () => {
+		const result = formatNumber(14, 44000, { dateNF: "dd/mm/yyyy" });
+		expect(result).toContain("/");
+	});
+
+	it("overrides 'm/d/yy' string with dateNF", () => {
+		const result = formatNumber("m/d/yy", 44000, { dateNF: "dd/mm/yyyy" });
+		// dateNF overrides the default "m/d/yy" format
+		expect(result).toContain("/");
+	});
+});
+
+describe("SSF: Chinese AM/PM detection", () => {
+	it("detects 上午/下午 as date format", () => {
+		expect(isDateFormat("\u4E0A\u5348/\u4E0B\u5348")).toBe(true);
+	});
+});
+
+describe("SSF: conditional format expressions", () => {
+	it("applies [>=1000] condition", () => {
+		const fmt = "[>=1000]#,##0;0";
+		const result = formatNumber(fmt, 1500);
+		expect(result).toContain(",");
+	});
+
+	it("falls through when condition not met", () => {
+		const fmt = "[>=1000]#,##0;0";
+		const result = formatNumber(fmt, 500);
+		expect(result).toBe("500");
+	});
+
+	it("applies [<0] condition", () => {
+		const fmt = "[<0](0);0";
+		const result = formatNumber(fmt, -5);
+		expect(result).toContain("(");
+	});
+
+	it("applies [=0] condition", () => {
+		const fmt = '[=0]"-";0';
+		const result = formatNumber(fmt, 0);
+		expect(result).toBe("-");
+	});
+
+	it("applies [<>0] condition", () => {
+		const fmt = '[<>0]0;"-"';
+		const result = formatNumber(fmt, 5);
+		expect(result).toBe("5");
+	});
+
+	it("applies [<=100] condition", () => {
+		const fmt = "[<=100]0.00;0.00";
+		const result = formatNumber(fmt, 50);
+		expect(result).toBe("50.00");
+	});
+
+	it("uses section 3 when both conditions have expressions and neither matches", () => {
+		const fmt = '[>100]0.00;[<0]0.00;"-"';
+		const result = formatNumber(fmt, 0);
+		expect(result).toBe("-");
+	});
+});
+
+describe("SSF: miscellaneous format branches", () => {
+	it("formats with General and non-number value", () => {
+		const result = formatNumber("General", true);
+		expect(result).toBe("TRUE");
+	});
+
+	it("formats with General and false", () => {
+		const result = formatNumber("General", false);
+		expect(result).toBe("FALSE");
+	});
+
+	it("formats undefined as empty string", () => {
+		const result = formatNumber("General", undefined);
+		expect(result).toBe("");
+	});
+
+	it("formats null as empty string", () => {
+		const result = formatNumber("General", null);
+		expect(result).toBe("");
+	});
+
+	it("formats text with @ placeholder", () => {
+		const result = formatNumber("@", "hello");
+		expect(result).toBe("hello");
+	});
+
+	it("4-section format: text section", () => {
+		const result = formatNumber("#,##0;-#,##0;0;@", "text");
+		expect(result).toBe("text");
+	});
+
+	it("formats with Infinity", () => {
+		const result = formatNumber("0.00", Infinity);
+		expect(typeof result).toBe("string");
+	});
+
+	it("formats with format table lookup by index", () => {
+		// Format 2 = "0.00" in the standard table
+		const result = formatNumber(2, 42);
+		expect(result).toBe("42.00");
+	});
+
+	it("formats with unknown index falls back to General", () => {
+		const result = formatNumber(999, 42);
+		expect(result).toBe("42");
+	});
+
+	it("formats with custom table option", () => {
+		const result = formatNumber(200, 42, { table: { 200: "0.0" } });
+		expect(result).toBe("42.0");
+	});
+});
+
+describe("SSF: write_num_flt specific patterns", () => {
+	it("handles #? pattern (question marks)", () => {
+		const result = formatNumber("??", 5);
+		expect(result.length).toBe(2);
+	});
+
+	it("handles large number in flr() path (>2^31)", () => {
+		const result = formatNumber("0", 3000000000);
+		expect(typeof result).toBe("string");
+	});
+
+	it("handles decimal with format 0.00", () => {
+		const result = formatNumber("0.00", 3.1);
+		expect(result).toBe("3.10");
+	});
+
+	it("handles decimal rounding carry", () => {
+		// 9.999 with format 0.00 should carry
+		const result = formatNumber("0.00", 9.999);
+		expect(result).toContain("10");
+	});
+});
+
+describe("SSF: integer-specific patterns", () => {
+	it("formats integer with leading zeros 00000", () => {
+		expect(formatNumber("00000", 42)).toBe("00042");
+	});
+
+	it("formats integer with fraction pattern", () => {
+		const result = formatNumber("# ?/?", 3);
+		expect(result).toContain("3");
+	});
+
+	it("formats integer with $ prefix", () => {
+		const result = formatNumber("$00", 42);
+		expect(result).toBe("$42");
+	});
+
+	it("formats integer with trailing comma (thousands scale)", () => {
+		const result = formatNumber("0,", 5000);
+		expect(result).toBe("5");
+	});
+});
+
+describe("SSF: parseExcelDateCode edge cases", () => {
+	it("parses serial 0 (Jan 0, 1900)", () => {
+		const result = parseExcelDateCode(0);
+		expect(result).toBeDefined();
+		expect(result!.day).toBe(0);
+		expect(result!.month).toBe(1);
+		expect(result!.year).toBe(1900);
+	});
+
+	it("parses serial 60 (Feb 29, 1900 — phantom)", () => {
+		const result = parseExcelDateCode(60);
+		expect(result).toBeDefined();
+		expect(result!.day).toBe(29);
+		expect(result!.month).toBe(2);
+		expect(result!.year).toBe(1900);
+	});
+
+	it("parses serial with sub-second overflow", () => {
+		// Time fraction just below next second — subSeconds should be normalized
+		const result = parseExcelDateCode(1.9999999);
+		expect(result).toBeDefined();
+		expect(result!.subSeconds).toBeGreaterThanOrEqual(0);
+	});
+
+	it("parses with Hijri mode", () => {
+		const result = parseExcelDateCode(44000, {}, true);
+		expect(result).toBeDefined();
+		// Hijri year should be much less than Gregorian
+		expect(result!.year).toBeLessThan(1900);
+	});
+
+	it("parses serial 0 with Hijri mode", () => {
+		const result = parseExcelDateCode(0, {}, true);
+		expect(result).toBeDefined();
+	});
+
+	it("returns null for negative serial", () => {
+		const result = parseExcelDateCode(-1);
+		expect(result).toBeNull();
+	});
+});
+
+describe("SSF: additional format edge cases", () => {
+	it("formats with escaped character \\x", () => {
+		const result = formatNumber("0\\x", 5);
+		expect(result).toContain("x");
+		expect(result).toContain("5");
+	});
+
+	it("formats with quoted string literal", () => {
+		const result = formatNumber('0" units"', 5);
+		expect(result).toBe("5 units");
+	});
+
+	it("formats with underscore padding character", () => {
+		const result = formatNumber("0_)", 5);
+		expect(result).toContain("5");
+		expect(result).toContain(" ");
+	});
+
+	it("formats with asterisk fill", () => {
+		const result = formatNumber("0*-", 5);
+		expect(result).toContain("5");
+	});
+
+	it("formats boolean TRUE with format", () => {
+		const result = formatNumber("0.00", true);
+		expect(result).toBe("TRUE");
+	});
+
+	it("formats boolean FALSE with format", () => {
+		const result = formatNumber("0.00", false);
+		expect(result).toBe("FALSE");
+	});
+
+	it("formats empty string returns empty", () => {
+		const result = formatNumber("0.00", "");
+		expect(result).toBe("");
+	});
+
+	it("formats null returns empty", () => {
+		const result = formatNumber("0.00", null);
+		expect(result).toBe("");
+	});
+});
+
+describe("SSF: AM/PM time formatting", () => {
+	it("formats with h:mm AM/PM", () => {
+		// 0.25 = 6:00 AM
+		const result = formatNumber("h:mm AM/PM", 0.25);
+		expect(result).toContain("AM");
+	});
+
+	it("formats with h:mm AM/PM (afternoon)", () => {
+		// 0.75 = 6:00 PM
+		const result = formatNumber("h:mm AM/PM", 0.75);
+		expect(result).toContain("PM");
+	});
+
+	it("formats with h:mm A/P", () => {
+		const result = formatNumber("h:mm A/P", 0.25);
+		expect(result).toContain("A");
+	});
+
+	it("formats with h:mm A/P (afternoon)", () => {
+		const result = formatNumber("h:mm A/P", 0.75);
+		expect(result).toContain("P");
+	});
+});
+
+describe("SSF: day-of-week and month-name formats", () => {
+	it("formats with ddd (abbreviated day)", () => {
+		const result = formatNumber("ddd", 44000);
+		expect(result.length).toBeLessThanOrEqual(3);
+	});
+
+	it("formats with dddd (full day name)", () => {
+		const result = formatNumber("dddd", 44000);
+		expect(result.length).toBeGreaterThan(3);
+	});
+
+	it("formats with mmm (abbreviated month)", () => {
+		const result = formatNumber("mmm", 44000);
+		expect(result.length).toBe(3);
+	});
+
+	it("formats with mmmm (full month name)", () => {
+		const result = formatNumber("mmmm", 44000);
+		expect(result.length).toBeGreaterThan(3);
+	});
+
+	it("formats with mmmmm (first letter of month)", () => {
+		const result = formatNumber("mmmmm", 44000);
+		expect(result.length).toBe(1);
+	});
+});
+
+describe("SSF: color and bracket tokens", () => {
+	it("formats with [Red] color prefix", () => {
+		const result = formatNumber("[Red]0.00", 42);
+		expect(result).toBe("42.00");
+	});
+
+	it("formats with [Blue] color prefix", () => {
+		const result = formatNumber("[Blue]0.00", 42);
+		expect(result).toBe("42.00");
+	});
+
+	it("formats with [Color5] numeric color", () => {
+		const result = formatNumber("[Color5]0.00", 42);
+		expect(result).toBe("42.00");
+	});
+});
+
+describe("ssf/format.ts — formatNumber", () => {
+	describe("General format", () => {
+		it("should format integer", () => {
+			expect(formatNumber("General", 42)).toBe("42");
+		});
+
+		it("should format decimal", () => {
+			expect(formatNumber("General", 3.14)).toBe("3.14");
+		});
+
+		it("should format very small number with E notation", () => {
+			// Need 1e-10 or smaller to trigger E notation in General format
+			const result = formatNumber("General", 1e-10);
+			expect(result).toContain("E");
+		});
+
+		it("should format very large number with E notation", () => {
+			const result = formatNumber("General", 1e15);
+			expect(result.length).toBeLessThan(20);
+		});
+
+		it("should format boolean true", () => {
+			expect(formatNumber("General", true)).toBe("TRUE");
+		});
+
+		it("should format boolean false", () => {
+			expect(formatNumber("General", false)).toBe("FALSE");
+		});
+
+		it("should format empty string", () => {
+			expect(formatNumber("General", "")).toBe("");
+		});
+
+		it("should format text string", () => {
+			expect(formatNumber("@", "hello")).toBe("hello");
+		});
+	});
+
+	describe("Number formats", () => {
+		it("should format #,##0.00", () => {
+			expect(formatNumber("#,##0.00", 1234.5)).toBe("1,234.50");
+		});
+
+		it("should format 0.00", () => {
+			expect(formatNumber("0.00", 3.1)).toBe("3.10");
+		});
+
+		it("should format zero with custom section", () => {
+			expect(formatNumber('0;0;"zero"', 0)).toBe("zero");
+		});
+	});
+
+	describe("Date formats", () => {
+		it("should format m/d/yy", () => {
+			const result = formatNumber("m/d/yy", 45292); // 2024-01-01
+			expect(result).toMatch(/1\/1\/24/);
+		});
+
+		it("should format yyyy-mm-dd", () => {
+			const result = formatNumber("yyyy-mm-dd", 45292);
+			expect(result).toBe("2024-01-01");
+		});
+
+		it("should format dd-mmm-yyyy", () => {
+			const result = formatNumber("dd-mmm-yyyy", 45292);
+			expect(result).toMatch(/01-Jan-2024/);
+		});
+
+		it("should format h:mm:ss", () => {
+			const result = formatNumber("h:mm:ss", 0.5); // noon
+			expect(result).toBe("12:00:00");
+		});
+
+		it("should format mm/dd/yyyy hh:mm:ss", () => {
+			const result = formatNumber("mm/dd/yyyy hh:mm:ss", 45292.5);
+			expect(result).toContain("01/01/2024");
+			expect(result).toContain("12:00:00");
+		});
+
+		it("should format elapsed hours [h]:mm", () => {
+			const result = formatNumber("[h]:mm", 1.5); // 36 hours
+			expect(result).toBe("36:00");
+		});
+
+		it("should format elapsed minutes [mm]:ss", () => {
+			const result = formatNumber("[mm]:ss", 0.5); // 720 minutes
+			expect(result).toBe("720:00");
+		});
+
+		it("should format sub-seconds", () => {
+			const result = formatNumber("h:mm:ss.00", 0.50001);
+			expect(result).toMatch(/12:00:00/);
+		});
+
+		it("should handle date serial 60 (1900 leap year bug)", () => {
+			const result = formatNumber("yyyy-mm-dd", 60);
+			expect(result).toBe("1900-02-29"); // phantom date
+		});
+
+		it("should handle date1904 option", () => {
+			const result = formatNumber("yyyy-mm-dd", 0, { date1904: true });
+			expect(result).toBe("1904-01-01");
+		});
+	});
+
+	describe("Scientific notation", () => {
+		it("should format 0.00E+00", () => {
+			const result = formatNumber("0.00E+00", 12345);
+			expect(result).toMatch(/1\.23E\+04/);
+		});
+
+		it("should format small number", () => {
+			const result = formatNumber("0.00E+00", 0.00123);
+			expect(result).toMatch(/E/);
+		});
+
+		it("should format negative scientific", () => {
+			const result = formatNumber("0.00E+00", -12345);
+			expect(result).toContain("-");
+		});
+	});
+
+	describe("Fraction formats", () => {
+		it("should format # ?/?", () => {
+			const result = formatNumber("# ?/?", 1.5);
+			expect(result.trim()).toContain("1/2");
+		});
+
+		it("should format # ??/??", () => {
+			const result = formatNumber("# ??/??", 3.333333);
+			expect(result.trim()).toMatch(/3\s+1\/\s*3/);
+		});
+
+		it("should format value < 1", () => {
+			const result = formatNumber("# ?/?", 0.5);
+			expect(result.trim()).toContain("1/2");
+		});
+	});
+
+	describe("Special formats", () => {
+		it("should handle currency [$]", () => {
+			const result = formatNumber("[$€-407]#,##0.00", 1234.5);
+			expect(result).toContain("€");
+			expect(result).toContain("1,234.50");
+		});
+
+		it("should handle text format @", () => {
+			expect(formatNumber("@", "hello")).toBe("hello");
+		});
+
+		it("should handle escaped characters", () => {
+			const result = formatNumber('0" kg"', 5);
+			expect(result).toBe("5 kg");
+		});
+
+		it("should return empty for null", () => {
+			expect(formatNumber("0", null)).toBe("");
+		});
+
+		it("should return #NUM! for NaN with number format", () => {
+			expect(formatNumber("0", NaN)).toBe("#NUM!");
+		});
+
+		it("should return #DIV/0! for Infinity with number format", () => {
+			expect(formatNumber("0", Infinity)).toBe("#DIV/0!");
+		});
+
+		it("should use format index lookup", () => {
+			// Format 14 = "m/d/yy"
+			const result = formatNumber(14, 45292);
+			expect(result).toMatch(/1\/1\/24/);
+		});
+
+		it("should use dateNF override for format 14", () => {
+			const result = formatNumber(14, 45292, { dateNF: "yyyy-mm-dd" });
+			expect(result).toBe("2024-01-01");
+		});
+
+		it("should use dateNF override for 'm/d/yy' string", () => {
+			const result = formatNumber("m/d/yy", 45292, { dateNF: "yyyy-mm-dd" });
+			expect(result).toBe("2024-01-01");
+		});
+
+		it("should handle two-section format", () => {
+			const result = formatNumber("0;0", -500);
+			expect(result).toBe("500");
+		});
+
+		it("should format text in 4-section format", () => {
+			const result = formatNumber("0;0;0;@", "text");
+			expect(result).toBe("text");
+		});
+
+		it("should format text with single section containing @", () => {
+			const result = formatNumber("@", "mytext");
+			expect(result).toBe("mytext");
+		});
+
+		it("should format Date object", () => {
+			const d = new Date("2024-01-01T00:00:00Z");
+			const result = formatNumber("yyyy-mm-dd", d);
+			// Date → serial number conversion depends on local timezone;
+			// just verify we get a valid yyyy-mm-dd date string back
+			expect(result).toMatch(/^\d{4}-\d{2}-\d{2}$/);
 		});
 	});
 });
